% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_bias.R
\name{AA_bias}
\alias{AA_bias}
\title{AA_bias}
\usage{
AA_bias(Ev, DB)
}
\arguments{
\item{Ev}{The evidence data frame ; must contain a "Sequence" column.}

\item{DB}{The search database, processed as a data.frame by Format.DB; must contain a "Sequence" column.}
}
\description{
A little function to see if there is a bias against certain amino acids in the dataset, versus parent search database.
}
\examples{
AA.bias <- AA_bias(Ev = ev, DB = db)

}
