% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrOrg.R
\name{abbrOrg}
\alias{abbrOrg}
\title{abbrOrg}
\usage{
abbrOrg(x, sep = ";")
}
\arguments{
\item{x}{String to abbreviate.}

\item{sep}{Name separator, default = ";"}
}
\value{
String of abbreviated names.
}
\description{
A utility function to abbreviate strings of Organism names.
 e.g. c("Homo sapiens;Mus musculus (C57Bl6)", "Escherichia coli") -> c("H. sapiens;M. musculus", "E. coli")
}
\examples{
x <- c("Homo sapiens;Mus musculus (C57Bl6)", "Escherichia coli")
abbrOrg(x)

}
