% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvNorm.IL.R
\name{AdvNorm.IL}
\alias{AdvNorm.IL}
\title{AdvNorm.IL}
\usage{
AdvNorm.IL(df, ids.col = "Unique State", exprs.col, exprs.log = FALSE, K = 5)
}
\arguments{
\item{df}{The input data frame, containing unique peptide identifiers and expression data.}

\item{ids.col}{The name of the column in df containing peptide identifiers. Default is "Unique State", a column I usually create which contains modified sequence pasted to charge.}

\item{exprs.col}{The name of the columns in df which contain expression values.}

\item{exprs.log}{Default = FALSE, meaning the input data is not log-transformed. Set instead to the base of the log if input data is log-transformed. If set to TRUE, the data is assumed to be log10 base. The data returned will be transformed or not as per the input.}

\item{K}{A fold factor defining the minimum (1/K) and maximum (*K) range of accepted values for correction factors.}
}
\value{
Normalized data. New column names are recycled from the original ones (in exprs.col), with "AdvNorm. " added as a prefix.
}
\description{
This function performs simplified "advanced normalization" on the columns of a data.frame.
}
\details{
This normalizes the data using the Levenberg Marquard procedure (aka. "advanced normalization") to minimize column-wise the sum square difference between different quantitative vectors.
The ".IL" suffix in the name stands for "isobaric labeling", because at first this was written with a TMT or iTRAQ reporter intensities data frame in mind.
However, this can be used for any data frame of quantitative values.
The input data.frame df must contain a column of IDs and 2 or more expression columns.
This is slower but is better than median only.
}
\examples{
adv.norm.data <- AdvNorm.IL(data, "Unique State", paste0("Reporter intensity ", c(0:9)), FALSE, K = 5)
require(ggplot2)
require(reshape2)
test <- melt(adv.norm.data)
test$value <- log10(test$value)
plot <- ggplot(test) + geom_density(stat = "density", aes(x = value, colour = variable))
grDevices::windows(width = 10, height = 10)
print(plot)

}
