% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvNorm.R
\name{AdvNorm}
\alias{AdvNorm}
\title{AdvNorm}
\usage{
AdvNorm(
  df,
  pep.ids.col = "Unique State",
  frac.col,
  exprs.col,
  exprs.log = FALSE
)
}
\arguments{
\item{df}{The input data frame, containing unique peptide identifiers, fraction numbers, and expression data.}

\item{pep.ids.col}{The name of the column in df containing peptide identifiers. Default is "Unique State", a column I usually create which contains modified sequence pasted to charge.}

\item{frac.col}{The name of the column in df which contains fraction numbers or raw file names.}

\item{exprs.col}{The name of the columns in df which contain expression values.}

\item{exprs.log}{Default = FALSE, meaning the input data is not log-transformed. Set instead to the base of the log if input data is log-transformed. If set to TRUE, assumes base 10. The data returned will be transformed or not as per the input.}
}
\description{
A function that performs advanced normalisation on dataframe containing a column of fractions names, peptide IDs, and 2 or more expression columns.
}
\examples{
adv.norm.data <- AdvNorm(df = data,
                         pep.ids.col = "Unique State",
                         frac.col = "Raw.file",
                         exprs.col = paste0("Reporter intensity ", c(0:9)),
                         exprs.log = FALSE)
require(ggplot2)
require(reshape2)
test <- melt(adv.norm.data)
test$value <- log10(test$value)
plot <- ggplot(test) + geom_density(stat = "density", aes(x = value, colour = variable))
grDevices::windows(width = 10, height = 10)
print(plot)
}
