% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaDIA_to_MQ.R
\name{alphaDIA_to_MQ}
\alias{alphaDIA_to_MQ}
\title{alphaDIA_to_MQ}
\usage{
alphaDIA_to_MQ(
  alphaDIA_fl,
  Fixed.mods = c("Carbamidomethyl"),
  N.clust,
  N.reserved = 1,
  cl,
  digPattern = "KR"
)
}
\arguments{
\item{alphaDIA_fl}{External alphaDIA precursors.tsv file.}

\item{Fixed.mods}{A vector of those Fixed modifications which will not be reported in the results (reported by alphaDIA, but MQ ignores them). Default = "Carbamidomethyl"}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{shinyOpt}{One of "popup" (default), "pane", "dialog", or "browser".}
}
\description{
Converts an alphaDIA precursors .tsv output table to a MaxQuant evidence.txt-like table.
As with the other functions of this type, the idea is not to get perfect conversion, but close enough that the data can be fed into this package's analysis scripts.

This will evolve as the Mann lab update https://alphadia.readthedocs.io/en/latest/methods/output-format.html
which as of 2025/07/16 is incomplete.

The output is a list of 2 items:
- Evidence: a data.frame similar to the evidence.txt table MaxQuant creates
- PTMs: a modifications table
}
\examples{
alphaDIA_fl <- rstudioapi::selectFile()
alphaDIA_fl <- Param$PSMs
temp <- alphaDIA_to_MQ(alphaDIA_fl)
ev <- temp$Evidences
Modifs <- temp$PTMs

}
