% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_from_GTF.R
\name{annot_from_GTF}
\alias{annot_from_GTF}
\title{annot_from_GTF}
\usage{
annot_from_GTF(file, CDS_only = TRUE, mode)
}
\arguments{
\item{file}{GFF or GTF (GFF v2) file.}

\item{CDS_only}{Logical: keep only CDS entries? This is a proteomics package so this is TRUE by default: this is usually all we care about.}

\item{mode}{One of "GFF" or "GTF". If missing, we will attempt to detect it from the file extension.}
}
\value{
A data.frame with parsed annotations.
}
\description{
Function to parse a GTF (or GFF) file to get annotations (for now only GO-terms).
}
\examples{
data <- annot_from_GTF(file)

}
