% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_to_tabl.R
\name{annot_to_tabl}
\alias{annot_to_tabl}
\title{annot_to_tabl}
\usage{
annot_to_tabl(
  x,
  start = "(",
  end = ")",
  Nterm = TRUE,
  Cterm = TRUE,
  sep = ";",
  numeric_data = FALSE
)
}
\arguments{
\item{x}{The annotated sequence character.}

\item{start}{The tag at the start of each annotation. Default = "("}

\item{end}{The tag at the end of each annotation. Default = ")"}

\item{Nterm}{Is there a symbol for the N-terminus? Default = TRUE}

\item{Cterm}{Is there a symbol for the C-terminus? Default = TRUE}

\item{sep}{Sometimes (for numeric annotations), there can be several per row, for instance when a peptide has several evidences. The values will be summarized (mean), so we need to know how they are separated. Default = ";"}

\item{numeric_data}{Is the data numeric? Default = FALSE}
}
\description{
A function to convert annotated sequences (MaxQuant format) into a table.
Useful for creating a table, of, e.g. sequence with amino-acid modifications, site probabilities, etc...
}
