% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Av_SE_fun.R
\name{Av_SE_fun}
\alias{Av_SE_fun}
\title{Av_SE_fun}
\usage{
Av_SE_fun(vect)
}
\arguments{
\item{vect}{Input numeric vector.}
}
\value{
Returns the arithmetic mean and standard error of a numeric vector.
}
\description{
A simple function to calculate the arithmetic mean and standard error of a numeric vector.
}
