% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.heatmap.R
\name{basic.heatmap}
\alias{basic.heatmap}
\title{basic.heatmap}
\usage{
basic.heatmap(
  matr,
  title = "Heatmap",
  subtitle = "",
  normRows = FALSE,
  isLog = FALSE,
  size = 2L,
  h.margins = c(5L, 10L),
  v.margin = 10L,
  colours = "D",
  colours_invert = FALSE,
  na.colour = "black",
  print = TRUE,
  save = FALSE,
  folder,
  return = FALSE,
  print_values = FALSE,
  hTree = TRUE,
  vTree = TRUE,
  split = FALSE
)
}
\arguments{
\item{matr}{The data frame or matrix to plot. If it has rownames or colnames, these will be displayed.}

\item{title}{The title of the graph.}

\item{normRows}{Default = FALSE. If TRUE, the data is re-normalized by row.}

\item{isLog}{Default = FALSE. Is the data log-transformed?}

\item{size}{The character size. Default = 2.5, you may want to reduce it for large proteins.}

\item{h.margins}{Optional: size of the horizontal margins. Default = c(5, 10)}

\item{v.margin}{Optional: size of the upper margin. Default = 3}

\item{colours}{Since October 2024, this should be a viridis scale name. The previous approach did not work.}

\item{colours_invert}{Default = FALSE. If TRUE, the viridis scale of choice is reverted.}

\item{na.colour}{Optional: What colour should NA values be displayed as?}

\item{print}{TRUE by default. But can be set to FALSE because sometimes you just want to save a plot for later without printing it.}

\item{save}{FALSE by default. Set to format you want pictures saved as ("jpeg", "png", "pdf", "svg"...)}

\item{folder}{Where to save the graph. Default = getwd()}

\item{return}{FALSE by default. Set to TRUE to return the created plot.}

\item{print_values}{If TRUE (default), values are printed on the heatmap.}

\item{hTree}{If TRUE (default), a horizontal (x) dendrogram will be plotted.}

\item{vTree}{If TRUE (default), a vertical (y) dendrogram will be plotted.}

\item{split}{Logical (default = FALSE), allows splitting labels over several lines.}
}
\value{
If return is TRUE, the evaluated ggplot is returned.
}
\description{
A function to create a simple heatmap of a data frame.
}
