% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNms2.R
\name{cleanNms3}
\alias{cleanNms3}
\title{cleanNms3}
\usage{
cleanNms3(
  names,
  rep = " ",
  experiments.map = Exp.map,
  aggregate.map = Aggregate.map,
  aggregate.list = Aggregate.list,
  RSA = Ref.Sample.Aggregate
)
}
\arguments{
\item{names}{The vector of sample names, or of column names containing sample names. Each name should be made of "___"-aggregated factor levels.}

\item{rep}{Replacement, user-friendlier separator. Default = " "}

\item{experiments.map}{The experiments map. Default = Exp.map}

\item{aggregate.map}{The aggregate map. Default = Aggregate.map}

\item{aggregate.list}{The named list of aggregates. Default = Aggregate.list}

\item{RSA}{List object describing individual samples and how they relate to the experimental factors. Default = Ref.Sample.Aggregate}
}
\description{
Meant as an internal function to clean up "___"-aggregated names for nicer plot annotations.
(NB: should work for other separators, we are just using "___" here).
For instance, it will turn "Exp1___Treated___KO___1" and "Exp1___Treated___KO" into the much user friendlier "Exp1 Treated KO 1" and "Exp1 Treated KO",
or, if - as is usually the case - there is only one Experiment, into "Treated KO 1" and "Treated KO".

Should be an improvement over cleanNms, which has the same purpose but a different code.
With this one it should be possible to feed a vector of data, example the colnames from a data.frame, and to fix all these names at once.

This is called cleanNms3 because cleanNms2 is a different function with a similar purpose used in the workflows for Venn Diagrams.
}
