% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustCheck.R
\name{clustCheck}
\alias{clustCheck}
\title{clustCheck}
\usage{
clustCheck(cl, n = N.clust, timeout = 5L)
}
\arguments{
\item{cl}{A cluster to check.}

\item{n}{Expected number of logical cores in the cluster.}

\item{timeout}{Default timeout = 5(s).}
}
\value{
TRUE if the cluster is working, otherwise FALSE.
}
\description{
Claude.AI-designed function to check our parallel cluster.
}
\examples{
tst <- clustCheck(parClust)

}
