% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configure.R
\name{Configure}
\alias{Configure}
\title{Configure}
\usage{
Configure()
}
\description{
Configuration function which will make sure some files from the package are moved to its subfolder in HOME upon first installation.
}
\examples{
aRmel::Configure()

}
