% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mod_seq.R
\name{cor_mod_seq}
\alias{cor_mod_seq}
\title{cor_mod_seq}
\usage{
cor_mod_seq(Pep, modseqcol = "Modified sequence", modcol = "Modifications")
}
\arguments{
\item{Pep}{A peptides table with a modified sequence column.}

\item{modseqcol}{The name of the modified sequence column. Default = "Modified sequence"}

\item{modcol}{The name of the modifications column. Only used if the file is actually in the old MaxQuant format. Default = "Modifications"}
}
\value{
A list of two objects:
- Peptides: the peptides table with (where applicable) modified sequence in 2-letters format
- PTMs: a modifications table
}
\description{
A function to edit recent MaxQuant verbose modified sequences to the old, more succinct two-letters format (pre MaxQuant 1.6.7.0)

If MaxQuant is from an older version, the modified sequences will not be modified, but the list of two objects will be created nonetheless.
}
\examples{
temp <- cor_mod_seq(ev)
# Check results:
View(temp$Peptides)
View(temp$PTMs)
# Assign results to discrete objects:
ev <- temp$Peptides
Modifs <- temp$PTMs

}
