% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Impute3.R
\name{Data_Impute3}
\alias{Data_Impute3}
\title{Data_Impute3}
\usage{
Data_Impute3(quant_data, groups, is.log = TRUE, seed)
}
\arguments{
\item{quant_data}{A matrix or data.frame with more than 1 column, containing missing or invalid (e.g. infinite).}

\item{groups}{A vector of groups of length equal to the number of columns of the input data frame.}

\item{is.log}{Default = TRUE. Are the input values log transformed? If set to FALSE, the data will be log-transformed prior to processing, then de-logged.}

\item{seed}{Use a fixed seed for reproducibility purposes.}
}
\description{
Work in Progress for a better wrapper function to impute missing values than my former versions, "Data_Impute" and "Data_Impute2" (the latter is the one used as of 29/06/2021).
Distinguishes between Missing-At-Random (MAR) and Missing-Not-At-Random, using knn for the former, QRILC for the latter.
The output is a list of the two arrays: new quantitative values and positions which were imputed.
}
\examples{
New_Data <- Data_Impute3(data)

}
