% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefArg.R
\name{DefArg}
\alias{DefArg}
\title{DefArg}
\usage{
DefArg(FUN, remove_NULLs = TRUE)
}
\arguments{
\item{FUN}{name of the function; can be entered as the function object or as its name (character).}

\item{remove_NULLs}{Should arguments with default value "NULL" be excluded?\cr
TRUE by default, because the risk is that some valuable variable in the parent environment can be "erased" as a result.}
}
\description{
Call this function for any function you want to troubleshoot.
This will create default values of the function's parameters in the main environment to allow easy debugging.

I used to have an in-house solution for this, but it was slightly buggy.
This is based on more standard R code and is more reliable.
}
\examples{
temp <- DefArg(F_test)

}
