% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensPlot.R
\name{DensPlot}
\alias{DensPlot}
\title{DensPlot}
\usage{
DensPlot(
  data,
  X,
  Series = NULL,
  Col = "black",
  Stat = "density",
  log = FALSE,
  print = TRUE,
  FUN = NULL,
  FUN.color = "red",
  title = "Density plot",
  save = FALSE,
  return = FALSE
)
}
\arguments{
\item{data}{The dataframe that contains the values.}

\item{X}{The name of the column of values whose density distribution is to be displayed.}

\item{Series}{The name of the column that maps values to series.}

\item{Col}{The series' colours. If there are more than one, must be a vector of colour names, named as series names, of the same length as the number of series. If unprovided default ggplot2 colours will be used instead.}

\item{Stat}{One of "count" or "density" (density).}

\item{log}{Should we log or not? If TRUE, the base is taken to be 10 by default; else you can instead specify the log.}

\item{print}{TRUE by default. But can be set to FALSE because sometimes you just want to save a plot for later without printing it.}

\item{FUN}{The optional function whose curve should be overlaid on the plot.}

\item{FUN.color}{The color of the curve of the optional function above. Default = "red".}

\item{title}{The title of the plot. Default = "Density plot"}

\item{save}{FALSE by default. Set to format you want pictures saved as.}

\item{return}{Should we return the plot? Default = FALSE.}
}
\description{
Just a small wrapper function for density plots, because I am tired of always typing the same stuff.
}
\examples{
Kolours <- colorRampPalette(c("blue", "red"))(length(unique(ev$Raw.file)))
names(Kolours) <- unique(ev$Raw.file)
DensPlot(ev, "Intensity", "Raw.file", Col = Kolours)

}
