% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANN_to_MQ.R
\name{DIANN_to_MQ}
\alias{DIANN_to_MQ}
\title{DIANN_to_MQ}
\usage{
DIANN_to_MQ(
  DIANN_fl,
  Fixed.mods = c("Carbamidomethyl"),
  PG.Genes = FALSE,
  Detect.Lib = TRUE,
  Min.Delta.Score = -Inf,
  N.clust,
  N.reserved = 1L,
  cl,
  shinyOpt = "popup",
  digPattern = "KR",
  log_Fl
)
}
\arguments{
\item{DIANN_fl}{External diaNN PSMs table file (.tsv, .parquet format, or both).}

\item{Fixed.mods}{A vector of those Fixed modifications which will not be reported in the results (reported by diaNN, but MQ ignores them). Default = "Carbamidomethyl"}

\item{PG.Genes}{If TRUE, will include per Protein Group/Gene quantitative and Q-values columns. Default = FALSE, because we will mostly be using this function in contexts where we want to perform protein groups inference re-calculate quantitative values ourselves.}

\item{Detect.Lib}{diaNN does not output M/Z values, but those are in the library generated at the end of some runs. If this is set to TRUE, it will try to detect the library and if available get M/Z values from there. Otherwise they will be computed.}

\item{Min.Delta.Score}{From which minimum delta score should we report a PSM? Default = -Inf (we assume that those few peptides with low delta score are because randomly the decoy database will include real-like peptides)}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{shinyOpt}{One of "popup" (default), "pane", "dialog", or "browser".}

\item{digPattern}{Default = "KR" for trypsin. Only used for the "Missed cleavages" column. Vector of amino acids after which the enzyme cleaves.}

\item{log_Fl}{If missing, the function will try to detect the log file based on the name of the report - but this may fail!}
}
\description{
Converts a diaNN PSMs table to a MaxQuant evidence.txt-like table.
As with the other functions of this type, the idea is not to get perfect conversion, but close enough that the data can be fed into this package's analysis scripts.

The output is a list of 3 items:
- Evidence: a data.frame similar to the evidence.txt table MaxQuant creates
- PTMs: a modifications table
- QuantUMS: was QuantUMS quantification on?

Surprisingly, the diaNN PSMs table does not include MZ values! Is there a way to do this by adding a command line argument?
For now they are either detected from the library file (same folder) or computed.
Cysteine alkylation is assumed to be Carbamidomethyl
}
\examples{
DIANN_fl <- rstudioapi::selectFile()
DIANN_fl <- Param$PSMs
temp <- DIANN_to_MQ(DIANN_fl)
ev <- temp$Evidences
Modifs <- temp$PTMs

}
