% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANN_to_MQ_server1.R
\name{DIANN_to_MQ_server1}
\alias{DIANN_to_MQ_server1}
\title{DIANN_to_MQ_server2}
\usage{
DIANN_to_MQ_server1(input, output, session)
}
\description{
Server for app used to decide whether to update PTM marks.
}
