% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANN_to_MQ_ui1.R
\name{DIANN_to_MQ_ui1}
\alias{DIANN_to_MQ_ui1}
\title{DIANN_to_MQ_ui1}
\usage{
DIANN_to_MQ_ui1(dflt = tstMap)
}
\description{
Function to return UI for app used to deciding whether to update PTM marks.
}
