% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffLog.R
\name{diffLog}
\alias{diffLog}
\title{diffLog}
\usage{
diffLog(p, dat)
}
\arguments{
\item{p}{Starting estimates, and N-1 length numeric vector.}

\item{dat}{Data.frame of log-transformed, row-normalized quantitative profiles with N columns.}
}
\description{
Internal function whose sole purpose is to be fed to another whose sum will be minimized by minpack.lm::nls.lm() within the LFQ.lm() function.
}
\examples{
temp3 <- ... # Data frame of log-transformed, row-normalized quantitative values
diffLog_v <- function(...) {
  p <- list(...)
  res <- diffLog(p, dat = temp3)
  return(res)
}

}
