% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Digest.R
\name{Digest}
\alias{Digest}
\title{Digest}
\usage{
Digest(
  Seq,
  Cut = c("K_", "R_"),
  strict.avoid = "",
  loose.avoid = c("K_P", "R_P"),
  missed = 2,
  min = 7,
  max = FALSE,
  characters.test = TRUE,
  AA = aRmel::AA,
  RemoveNtermMet = "loose",
  N.clust,
  N.reserved = 1,
  collapse,
  ChnkSz = Inf,
  cl
)
}
\arguments{
\item{Seq}{The character vector of protein Sequence(s) to digest.}

\item{Cut}{The character vector of amino acid(s) after which to cut. For each, provide one letter followed or preceded by underscore indicating whether the cut is C- or N-terminal, respectively. Default (for Trypsin) = c("K_", "R_")}

\item{strict.avoid}{The pattern(s) that will strictly block cleavage when overlapping with cleavage sites. Write cleavage site as underscore. Default = ""}

\item{loose.avoid}{The pattern(s) that will partially block cleavage when overlapping with cleavage sites. Write cleavage site as underscore. Default =  c("K_P", "R_P")}

\item{missed}{How many missed cleavages should be considered? Default = 2.}

\item{min}{The minimum length of peptides to report. Default = 7}

\item{max}{The maximum length of peptides to report. Set to FALSE (default) to allow any peptide length above min.}

\item{characters.test}{Should the compatibility of Sequences with the vector of valid amino acids be tested (True by default)?}

\item{AA}{Character vector of which amino Acids letters are considered valid? Default is the 20 standard ones plus S for selenocysteine: c("A","C","D","E","F","G","H","I","K","L","M","N","P","Q","R","S","T","U","V","W","Y")}

\item{RemoveNtermMet}{The default "loose" allows for both retention and loss of each N-terminal methionine. Set to "strict" to force loss. Set to "predict" to apply the rules as summarized in Wingfield PT. N-Terminal Methionine Processing. Curr Protoc Protein Sci. 2017 Apr 3;88:6.14.1-6.14.3. doi: 10.1002/cpps.29. PMID: 28369664; PMCID: PMC5663234.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1. Only used if missed > 0}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one. Only used if missed > 0}

\item{collapse}{For compatibility with older usage: option to specify a character (e.g. ";") to collapse the peptides list for each protein into a single character vector.}

\item{ChnkSz}{Deprecated}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}
}
\description{
A function to digest unmodified protein Sequence into peptides.

Will by default allow for loss of N-terminal methionine.
Default parameters digest like trypsin.
Proline after the R or K is assumed to block some but not all cleavages, hence both missed and cleaved peptides are included.

Uses parallel processing by default for now.
Still needs some speed boosts. The whole thing should be parallelized in one big go, as opposed to by bits.
The regex approach currently used may not be ideal. It provided a small speed bump when introduced to replace list-based approaches, but may more less safe.

Used by ProtMatch, but not by the much faster/better ProtMatch2.
}
\examples{
Digest(Seq = "MAAAAAAAAAKCCCCCCCRD")

}
