% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_worker.R
\name{.bind_worker}
\alias{.bind_worker}
\title{.bind_worker}
\usage{
.bind_worker(fun, data = list(), parent = baseenv())
}
\arguments{
\item{fun}{Function to bind}

\item{data}{Named list of variables used by the function}

\item{envir}{Environment in which to bind}
}
\description{
Worker function to bind variables to the environment of a function which is to be run in parallel within a larger function.
}
