% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.DB_worker.R
\name{.Format.DB_worker}
\alias{.Format.DB_worker}
\title{.Format.DB_worker}
\usage{
.Format.DB_worker(btch, Roolz, mode, IDs_only, trimName)
}
\arguments{
\item{btch}{Batch of fasta entries}

\item{Roolz}{List of regex rules}

\item{mode}{Mode string}

\item{IDs_only}{Logical, return only IDs?}

\item{trimName}{Logical, whether to trim names (anything after the last underscore will be trimmed)}
}
\description{
Worker function for Format.DB()
}
