% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP2MQ_modSeqWrkr1.R
\name{.FP2MQ_modSeqWrkr1}
\alias{.FP2MQ_modSeqWrkr1}
\title{.FP2MQ_modSeqWrkr1}
\usage{
.FP2MQ_modSeqWrkr1(x, mods, pat)
}
\arguments{
\item{x}{List, comma-split assigned modifications column from FragPipe.}

\item{mods}{Modifications table.}

\item{pat}{Regex for amino acids.}
}
\description{
Worker function used by FP_to_MQ().
}
