% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP2MQ_modSeqWrkr2.R
\name{.FP2MQ_modSeqWrkr2}
\alias{.FP2MQ_modSeqWrkr2}
\title{.FP2MQ_modSeqWrkr2}
\usage{
.FP2MQ_modSeqWrkr2(x, openSearch = OpenSearch)
}
\arguments{
\item{x}{Temp data, see FP_to_MQ code...}

\item{openSearch}{Is this an open search?}
}
\description{
Worker function used by FP_to_MQ().
}
