% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP2MQ_modSeqWrkr5.R
\name{.FP2MQ_modSeqWrkr5}
\alias{.FP2MQ_modSeqWrkr5}
\title{.FP2MQ_modSeqWrkr5}
\usage{
.FP2MQ_modSeqWrkr5(x, mods)
}
\arguments{
\item{x}{Temp data, see FP_to_MQ code...}

\item{mods}{Modifications table.}
}
\description{
Worker function used by FP_to_MQ().
}
