% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSave_worker.R
\name{.ggSave_worker}
\alias{.ggSave_worker}
\title{.ggSave_worker}
\usage{
.ggSave_worker(x)
}
\arguments{
\item{x}{A list with 4 slots:\cr
- Plot: the evaluated ggplot to save\cr
- Path: the directory in which to save the plot\cr
- Ttl: name of the destination file (without extension)\cr
- Ext: extension of the destination file}
}
\description{
Worker function within Volcano.plot() for saving evaluated ggplots.
}
