% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelEdit_worker.R
\name{.labelEdit_worker}
\alias{.labelEdit_worker}
\title{.labelEdit_worker}
\usage{
.labelEdit_worker(x, colchar = 25)
}
\arguments{
\item{x}{Character label split at spaces}

\item{colchar}{How many characters wide a column of text may be. Default = 25}
}
\description{
Worker function within Volcano.plot() for editing labels to make them multi-row.
}
