% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Rank_OR_Prof.R
\name{.plot_Rank_OR_Prof}
\alias{.plot_Rank_OR_Prof}
\title{.plot_Rank_OR_Prof}
\usage{
.plot_Rank_OR_Prof(ii)
}
\arguments{
\item{ii}{Integer, index of the row of the samplesDF controlling what version of the plot is drawn.}
}
\description{
Worker function used by the profile_and_rankedAbund_plots.R source to draw profile and ranked abundance plots.
Allows parallel drawing of all of these plots.
}
