% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySearch.R
\name{.plotlySearch}
\alias{.plotlySearch}
\title{.plotlySearch}
\usage{
.plotlySearch(p, search_data, fade_opacity = 0.2, placeholder = "Search...")
}
\arguments{
\item{p}{Plotly scatter plot (built or not) which should receive a search field.}

\item{search_data}{A data.frame derived from the data used to create the trace(s) of interest, with the same rows (number and ordering) and searchable columns.}

\item{fade_opacity}{Opacity of non-selected points. Default = 0.2.}

\item{placeholder}{Text displayed in the search field, default = "Search..."}
}
\value{
A build plotly widget with the search field added.
}
\description{
This function adds a search field to a plotly volcano plot.
It is NOT my work: it was designed under my supervision by chatGPT and istaGPT (our local GPT).

Note that this function has special behaviour for lightgrey points (rgba = 211, 211, 211, 1): matching points are set to rgba = 0, 0, 0, 1 while non matching points are set to rgba = 211, 211, 211, 0.2!
This allows seeing contrast between selected/non-selected points for the grey ("boring") area of volcano plots, where we use "lightgrey'.
}
\examples{
plotLy <- do.call(plotly::plot_ly, args_ly_)
plotLy <- .plotlySearch(plotLy, searchDat)

}
