% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyConfig.R
\name{.pyConfig}
\alias{.pyConfig}
\title{.pyConfig}
\usage{
.pyConfig()
}
\description{
A function to automatically check our python configuration, upgrade pip, install sdrf-pipelines and check that python scripts is in path.
}
