% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK2MQ_modSeqWrkr1.R
\name{.SK2MQ_modSeqWrkr1}
\alias{.SK2MQ_modSeqWrkr1}
\title{.SK2MQ_modSeqWrkr1}
\usage{
.SK2MQ_modSeqWrkr1(x, tempMods)
}
\arguments{
\item{x}{Temporary modified sequence}

\item{tempMods}{Temporary modifications table}
}
\description{
Worker function used by Skyline_to_MQ().
}
