% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK2MQ_modSeqWrkr2.R
\name{.SK2MQ_modSeqWrkr2}
\alias{.SK2MQ_modSeqWrkr2}
\title{.SK2MQ_modSeqWrkr2}
\usage{
.SK2MQ_modSeqWrkr2(x)
}
\arguments{
\item{x}{Temporary modified sequence}
}
\description{
Worker function used by Skyline_to_MQ().
}
