% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWrkr1.R
\name{.summaryWrkr1}
\alias{.summaryWrkr1}
\title{.summaryWrkr1}
\usage{
.summaryWrkr1(x, raw, type)
}
\arguments{
\item{x}{File.}

\item{raw}{Table of MS raw files.}

\item{type}{Type of chromatogram to export.}
}
\description{
Worker function used by MQ_summary().
}
