% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWrkr2.R
\name{.summaryWrkr2}
\alias{.summaryWrkr2}
\title{.summaryWrkr2}
\usage{
.summaryWrkr2(x, tempDat, size)
}
\arguments{
\item{x}{Raw file}

\item{tempDat}{Temp data.}

\item{size}{Type of chromatogram to export.}
}
\description{
Worker function used by MQ_summary().
}
