% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.DB.R
\name{Format.DB}
\alias{Format.DB}
\title{Format.DB}
\usage{
Format.DB(
  file,
  mode = "UniProt",
  in.env = FALSE,
  Full.ID.rule = "^>([^ ]*)",
  Protein.ID.rule = "^>[a-z]{2}\\\\|([^\\\\|]*)",
  Name.rule = "^>[a-z]{2}\\\\|[^\\\\|]*\\\\|([^_]*)",
  Unique = TRUE,
  IDs_only = FALSE,
  parallel = FALSE,
  N.clust,
  N.reserved = 1,
  cl
)
}
\arguments{
\item{file}{Name (as character) of the Fasta file.}

\item{mode}{One of "Uniprot", "Ensembl", "RefSeq-RNA", "RefSeq-Protein", "RefSeq-CDS", "NCBI", "TAIR" or "custom".}

\item{in.env}{If TRUE, the function will attempt processing the character string in one of two ways: a) if it recognizes the character string as a Fasta file, it will directly process it; else, it will treat it as the name of a Fasta file already in the environment. If set to FALSE (default), it will attempt to load the file from the work directory.}

\item{Full.ID.rule}{Regex rule used to parse the header and extract the full ID. Only used if operating in custom mode. Default is "^>([^ ]*)" (UniProt).}

\item{Protein.ID.rule}{Regex rule used to parse the header and extract protein ID. Only used if operating in custom mode. Default is "^>[a-z]{2}\\|([^\\|]*)" (UniProt).}

\item{Name.rule}{Regex rule used to parse the header and extract the protein name. Only used if operating in custom mode. Default is "^>[a-z]{2}\\|[^\\|]*\\|([^_]*)" (UniProt).}

\item{Unique}{If TRUE (default), will filter fasta to provide only one accession per sequence. (This may mean that only 1 gene will be retained where several encode the exact same protein, but reducing redundancy is good for FDR.)}

\item{IDs_only}{Allows extracting just the IDs (default = FALSE). Used internally only.}

\item{parallel}{Defult = FALSE; if TRUE, uses parallel processing and the next few arguments.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}
}
\description{
A function to load Fasta databases from the work directory or the environment and format them into a data frame.
Depending on the mode you select, specific regex rules are applied to the header to extract information.
Additionally, three rules can be custom specified by the arguments.
Now parallelized!!!
}
\examples{
db <- Format.DB(file = "Search database.FASTA", in.env = TRUE, species = "HUMAN")
db <- Format.DB(file = "Search database2.FASTA", in.env = FALSE, species = "RAT;MOUSE")
CDS <- Format.DB(list.files()[grep(".*\\\\.CDS\\\\..*\\\\.FASTA$", toupper(list.files()))], mode = "ensembl")

}
