% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.DB_txt.R
\name{Format.DB_txt}
\alias{Format.DB_txt}
\title{Format.DB_txt}
\usage{
Format.DB_txt(
  txt,
  filter = NULL,
  GO = TRUE,
  Taxonomy = TRUE,
  InterPro = TRUE,
  Pfam = TRUE,
  PIRSF = TRUE,
  PROSITE = TRUE,
  EMBL = TRUE,
  Ensembl = TRUE,
  MW = TRUE,
  Sequence = TRUE,
  PTMs = FALSE,
  PDB = TRUE,
  TAIR = TRUE,
  WormBase = TRUE,
  FlyBase = TRUE,
  Features = FALSE,
  usePar = TRUE,
  N.clust,
  N.reserved = 1,
  cl,
  Feat_accRgx = "[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}"
)
}
\arguments{
\item{txt}{Text annotations, normally already read into R from your UniProtKB annotations txt file, but it is now also possible to provide a path to a file.}

\item{filter}{Valid UniProt Accessions. Filter for these to hasten the process.}

\item{GO}{Set to TRUE (default) to retrieve GO annotations.}

\item{Taxonomy}{Set to TRUE (default) to retrieve taxonomic annotations.}

\item{InterPro}{Set to TRUE (default) to retrieve InterPro annotations.}

\item{Pfam}{Set to TRUE (default) to retrieve Pfam annotations.}

\item{PIRSF}{Set to TRUE (default) to retrieve PIRSF annotations.}

\item{PROSITE}{Set to TRUE (default) to retrieve PROSITE annotations.}

\item{EMBL}{Set to TRUE (default) to retrieve EMBL annotations.}

\item{Ensembl}{Set to TRUE (default) to retrieve Ensembl annotations.}

\item{MW}{Set to TRUE (default) to retrieve protein molecular weights.}

\item{Sequence}{Set to TRUE (default) to retrieve protein sequences.}

\item{PTMs}{Default = FALSE; if TRUE, will return a map with all PTM sites for which there is evidence.}

\item{PDB}{Set to TRUE (default) to retrieve PDB IDs.}

\item{TAIR}{Set to TRUE (default) to retrieve TAIR IDs.}

\item{WormBase}{Set to TRUE (default) to retrieve WormBase IDs.}

\item{FlyBase}{Set to TRUE (default) to retrieve FlyBase IDs.}

\item{Features}{Set to TRUE (default) to retrieve Features.}

\item{usePar}{Default = FALSE; if TRUE, uses the next 3 arguments and the "parallel" package for multi-threads parallel processing.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{Feat_accRgx}{Normally do not change this, the default regex from https://www.uniprot.org/help/accession_numbers is fine.\cr
This is only used if Features = TRUE, and has NO EFFECT on the Accession column itself!\cr
Only provide a different regex here if you are parsing a custom file not sourced from UniProtKB where the Accession pattern changes.}
}
\description{
Function to get a table with all entry details from a text version of a UniProtKB proteome.
(download all protein entries and choose "text" format)
}
\examples{
txt <- readLines("D:/Fasta_databases/Arabidopsis_thaliana/Arabidopsis_thaliana_UP_20230202_Iso.txt")
data <- Format.DB_txt(txt)

}
