% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Mod_Pep.R
\name{Format_Mod_Pep}
\alias{Format_Mod_Pep}
\title{Format_Mod_Pep}
\usage{
Format_Mod_Pep(x)
}
\arguments{
\item{x}{The Modified sequences from the evidence file.}
}
\description{
A function to correct the evidences table's Modified sequence file, because since MaxQuant 1.6.?.? a new format is being used which prevents my previous scripts from working.
Will output the corrected modified sequences as well as a modifications table for inspection.
This is a duplicate, currently unused, instead I use the more current cor_mod_seq: I appear to have created 2 functions for the same purpose.
Candidate for deletion.
}
\examples{
test <- Format_Mod_Pep(ev$Modified.sequence)
View(test$"Modifications table")
ev$Modified.sequence <- test$"Corrected sequences"

}
