% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_to_MQ.R
\name{FP_to_MQ}
\alias{FP_to_MQ}
\title{FP_to_MQ}
\usage{
FP_to_MQ(
  FP_Workflow,
  FP_Manifest,
  Min.Delta.Score = -Inf,
  UniProtIDs = TRUE,
  FailIfNoQuant = FALSE,
  N.clust,
  N.reserved = 1,
  cl,
  OpenSearch = FALSE
)
}
\arguments{
\item{FP_Workflow}{Fragpipe workflow. Parsed to identify FragPipe's output directory, search parameters, incl. modifications.}

\item{FP_Manifest}{FragPipe's samples manifest, required for mapping raw files to samples ("Experiments" in the MaxQuant sense).}

\item{Min.Delta.Score}{From which minimum delta score should we report a PSM? Default = -Inf (we assume that those few peptides with low delta score are because randomly the decoy database will include real-like peptides)}

\item{UniProtIDs}{If TRUE (default), long protein IDs are re-processed to only get the accession.}

\item{FailIfNoQuant}{If TRUE, the function will fail with an error if it does not detect quantitative values (which, in FragPipe, are optional). It is useful in most case (otherwise a workflow may fail further down the road), but I am setting this to FALSE by default for now to avoid stupid fails for ident-only workflows.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{OpenSearch}{Is it an open search? If TRUE, will also ignore Open Search-type columns even if they are present. If TRUE, will try to detect and process them - but this is probably still a bit buggy.}
}
\value{
A list of (at least) 4 items:
- "Evidence": a PSMs table similar (nor identical!) to the evidence table MaxQuant creates
- "PTMs": a modifications table
- "FracMap": a map of the different MS raw files and their relationship to Experiments
- "WorkFlow": the FragPipe workflow
}
\description{
Reads FragPipe's workflow and manifest files
From them, detects and parses psm.tsv tables (one per Experiment) to a single MaxQuant evidence.txt-like table.
As with the other functions of this type, the idea is not to get perfect conversion, but close enough that the data can be fed into this package's analysis scripts.
}
\examples{
FP.Wrkflw <- paste0(wd, "/fragpipe.workflow")
FP.Mnfst <- paste0(wd, "/fragpipe-files.fp-manifest")
temp <- FP_to_MQ(FP.Wrkflw, FP.Mnfst)
ev <- temp$Evidences
Modifs <- temp$PTMs

}
