% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{geomean}
\usage{
geomean(x, filter = TRUE)
}
\arguments{
\item{x}{The numeric data which will be averaged.}

\item{filter}{If set to TRUE (default), this will remove all non numeric, NA, NaN and +/-Inf values.}
}
\description{
Calculates the geometric mean of linear-scale data.
NB: I don't think that this function is being used anymore, is it?
}
\examples{
> geomean(c(1, 3))
  1.732051
  
}
