% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistones.R
\name{getHistones}
\alias{getHistones}
\title{getHistones}
\usage{
getHistones(DB)
}
\arguments{
\item{DB}{A data.frame with at least "Header" and "Protein ID" columns.}
}
\value{
A list of 2 vectors:
 - All = accessions of all histones
 - Core = accessions of core histones
}
\description{
This function takes a data.frame resulting from the parsing of a fasta file by Format.DB() and text mines its header to identify histones.
}
\examples{
histIDs <- getHistones(db)

}
