% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_annotate2.R
\name{GO_annotate2}
\alias{GO_annotate2}
\title{GO_annotate2}
\usage{
GO_annotate2(
  proteins,
  species,
  chunks = 6,
  tries = 5,
  kegg = FALSE,
  reactome = FALSE
)
}
\arguments{
\item{proteins}{The character vector of UniProt protein IDs to annotate.}

\item{species}{The species of the proteins.}

\item{chunks}{Optional, but set by default to 6. Size of the chunks to attempt to download. Emergency argument, normally data should be querried in one batch but this currently fails. May be set to missing by default in future.}

\item{tries}{When working by chunks, how many times should I attempt each chunk before skipping?}

\item{kegg}{Default = FALSE. Should KEGG annotations be downloaded too?}

\item{reactome}{Default = FALSE. Should Reactome annotations be downloaded too?}
}
\description{
A function to annotate a vector of proteins with GO terms. Can also do KEGG and Reactome pathways, despite its more restrictive name.
This was meant to be an improved version of the previous one, which did not work well at all...
but I think does not work now, at least I do not think it is being used.
Candidate for deletion.

Returns the vector of protein IDs as data.frame with added GO columns.
}
\examples{
test <- GO_annotate2(db$Protein.ID, "Homo sapiens")

}
