% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_annotate_DB.R
\name{GO_annotate_DB}
\alias{GO_annotate_DB}
\title{GO_annotate_DB}
\usage{
GO_annotate_DB(proteins, species, type = "Standard")
}
\arguments{
\item{proteins}{The character vector of UniProt protein IDs to annotate.}

\item{species}{The species of the proteins.}

\item{type}{The type of organism.}
}
\description{
A function to annotate a vector of proteins with GO terms.
}
\examples{
annot <- GO_annotate_DB(db$Protein.ID, "Homo sapiens")

}
