% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_annotate_PG.R
\name{GO_annotate_PG}
\alias{GO_annotate_PG}
\title{GO_annotate_PG}
\usage{
GO_annotate_PG(Prot, annotations, ids = "Protein IDs")
}
\arguments{
\item{Prot}{The protein groups file.}

\item{annotations}{The annotations file; must be the output of the GO_annotate_DB function.}

\item{ids}{The name of the column with the protein IDs; usually "Protein IDs" (default).}
}
\description{
A function to process the results of the GO_annotate_DB function into new columns added to a protein groups file.
}
\examples{
annot <- GO_annotate_DB(db$Protein.ID, "Homo sapiens")
GO_annotate_PG(PG, annotations = annot)

}
