% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsep.R
\name{grsep}
\alias{grsep}
\title{grsep}
\usage{
grsep(
  pattern,
  replacement = "",
  x,
  sep = ";",
  mode = "grep",
  ignore.case = FALSE,
  perl = FALSE,
  value = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  invert = FALSE
)
}
\arguments{
\item{pattern}{The IDs to search for.}

\item{replacement}{The replacement, only used if mode = gsub.}

\item{x}{A character vector listing IDs in which to search.}

\item{sep}{The characters list's separator. Default = ";"}

\item{mode}{Should this work in grep, gsub or grepl mode?}

\item{ignore.case}{The base parameter (default = FALSE)}

\item{perl}{The base parameter (default = FALSE)}

\item{value}{The base parameter (default = FALSE)}

\item{fixed}{The base parameter (default = FALSE)}

\item{useBytes}{The base parameter (default = FALSE)}

\item{invert}{The base parameter (default = FALSE)}
}
\description{
A wrapper function for grep/grepl/gsub to search IDs in a separated characters vector listing a set of IDs.
N.B.: Other modes could easily be added to the code without major rewrites of this or other scripts using it.
}
\examples{
tst <- grsep(c("MYFAVOURITEPROTEINSACCESSION", "ALSOTHISONE", "ANDHERESANOTHERONEFORGOODMEASURE"), x = PG$"Protein IDs")

}
