% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsep2.R
\name{grsep2}
\alias{grsep2}
\title{grsep2}
\usage{
grsep2(IDs, x, sep = ";", ignore.case = FALSE, value = FALSE, invert = FALSE)
}
\arguments{
\item{IDs}{The IDs to search for.}

\item{x}{A character vector listing IDs in which to search.}

\item{sep}{The characters list's separator. Default = ";"}

\item{ignore.case}{The base parameter (default = FALSE)}

\item{value}{Return values from matches (TRUE) or just match ids (FALSE, default)}

\item{invert}{The base parameter (default = FALSE)}
}
\description{
A function to look for IDs in a separated character list.
Unlike "grsep", it is not a wrapper for "grep/grepl". Unlike it, it should not break for large number of IDs.
Also unlike grsep it does not have "gsub-like" capabilities (yet?)
This function is based on reshape2; it is significantly faster than grsep on large vectors, but can be slower on smaller ones.
}
\examples{
tst <- grsep(c("MYFAVOURITEPROTEINSACCESSION", "ALSOTHISONE", "ANDHERESANOTHERONEFORGOODMEASURE"), PG$"Protein IDs")

}
