% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_to_seg.R
\name{hclust_to_seg}
\alias{hclust_to_seg}
\title{hclust_to_seg}
\usage{
hclust_to_seg(hc)
}
\arguments{
\item{hc}{A hclust object.}
}
\description{
Great function built with the help of chatGPT to extract segments and labels from an hclust efficiently, for the purpose of plotting them in a heatmap.
}
