% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertElems.R
\name{insertElems}
\alias{insertElems}
\title{insertElems}
\usage{
insertElems(vect, pos, elems, warn = FALSE)
}
\arguments{
\item{vect}{The target vector of length N.}

\item{pos}{The k positions where k elements should be inserted; pos can range from 0 (before the first item) to N (after the last).}

\item{elems}{The k elements to insert.}

\item{warn}{Should we display a warning when we have no choice but to convert a vector from numeric to character? Default = FALSE.}
}
\description{
Nifty little function to insert elements in a vector after the indicated positions.
}
\examples{
> insertElems(c(1:4), 3, "A")
[1] "1" "2" "A" "3" "4"

}
