% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.all.good.R
\name{is.all.good}
\alias{is.all.good}
\title{is.all.good}
\usage{
is.all.good(x, mode = "values")
}
\arguments{
\item{x}{Input numeric or integer vector, or all-numeric/integer matrix.}

\item{mode}{The function can work in two ways; if set to "values", "1" or 1 (default), removes all +Inf, -Inf, NA and NaN values. If set to "logical", "2" or 2, returns TRUE for real finite values and FALSE for others. If input is non-numeric, will throw a warning and attempt to convert to numeric.}
}
\description{
A function to filter numeric vectors to just include real numbers.
}
\examples{
is.all.good(c(1, NA, NaN, 2, -Inf, +Inf))
[1] 1 2
is.all.good(c(1, NA, NaN, 2, -Inf, +Inf), mode = "logical")
[1]  TRUE FALSE FALSE  TRUE FALSE FALSE
}
