% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFQ.lm.R
\name{LFQ.lm}
\alias{LFQ.lm}
\title{LFQ.lm}
\usage{
LFQ.lm(
  ids,
  InputTabl,
  id = "id",
  IntensCol,
  Summary.method = "median",
  Summary.weights,
  Min.N = 2,
  Max.N = 50,
  Viz = FALSE,
  Is.log = TRUE
)
}
\arguments{
\item{ids}{Integer, a vector/list of ids in InputTabl to summarize.}

\item{InputTabl}{Table containing individual entities (peptides or fragments) to summarize.}

\item{id}{Name of the IDs column in InputTabl. Default = "id"}

\item{IntensCol}{Names of the quantitative columns in InputTabl}

\item{Summary.method}{The summary method used for ratios (the Levenberg-Marquardt algorithm is used for intensities). One of "mean", "median", or "weighted.mean".}

\item{Summary.weights}{If a "weighted.mean" summary method is chosen, then a vector of weights must be provided. This should be the name of a column of the peptides table}

\item{Min.N}{How many peptides should at least be present? Should be at the very least 1 (default = 2).}

\item{Max.N}{How many peptides can we use at most for the Levenberg-Marquardt procedure? Using too many peptides can be an issue, e.g. with huge proteins like Titin. Default = 50. The most intense peptides will be selected.}

\item{Viz}{Should we plot the results? For testing only, default = FALSE.}

\item{Is.log}{Are input intensities log-transformed? TRUE by default.}

\item{Intensity.weights}{default = FALSE; if TRUE, will take into account individual peptide intensities when calculating average profile if "Summary.method" is either mean "mean" or "weighted.mean" (thus, it will actually be a weighted mean regardless).}
}
\description{
A function which summarizes individual peptide/fragment quantitative values into a single quantitative vector, using the Levenberg-Marquardt procedure.
Used at two levels:
- In the optional MS2corr2MS1.R source, to summarize a peptidoform's MS1 (precursor) and MS2 (fragment) measurements into a single quantitative vector.
- In Prot.Quant, to summarize peptidoform-level information into a protein group-levels quantitative vector.
The output is log10-transformed.
}
\examples{
lfq <- LFQ.lm(ids,
              InputTabl = Pep,
              IntensCol = Pep.Intens.Nms,
              Summary.method = Summary.method,
              Summary.weights = Summary.weights,
              Min.N = Min.N,
              Max.N = Max.N,
              Is.log = TRUE)
# Used within Prot.Quant

}
