% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Bckp.R
\name{load_Bckp}
\alias{load_Bckp}
\title{load_Bckp}
\usage{
load_Bckp(backup, startDir, clean = TRUE, loadPack = TRUE)
}
\arguments{
\item{startDir}{Here you can specify the directory in which the selection window will start.}

\item{clean}{If set to TRUE (default), will first remove all objects existing in the current environment.}

\item{loadPack}{Should we load packages? Default = TRUE}

\item{bckp}{The backup file to reload. If missing, opens a popup selection window.}
}
\description{
This function:
 - cleans the existing environment,
 - reloads a data analysis backup,
 - sets the proper working directory,
 - loads all necessary packages and
 - (if necessary) re-creates the original parallel cluster.
}
