% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Context.R
\name{load_Context}
\alias{load_Context}
\title{load_Context}
\usage{
load_Context(record, startDir, clean = FALSE)
}
\arguments{
\item{record}{The dataset details record .txt file to parse.}

\item{startDir}{In case record is not provided, you can specify here the directory in which the selection window will start. Ignored otherwise.}

\item{clean}{If set to TRUE, will first remove all objects existing in the current environment.}
}
\description{
This function takes a dataset details record file (as created when we run the Start_analysis.R app), then it reloads in environment:\cr
 - dataset name\cr
 - input and output directory\cr
 - temporary work directory\cr
and if the latter exists, sets work directory to it.
}
