% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFun.R
\name{loadFun}
\alias{loadFun}
\title{loadFun}
\usage{
loadFun(file, tryClassic = FALSE)
}
\arguments{
\item{file}{The path/file to save to.}

\item{tryClassic}{In past versions, we used the classic R save/save.image/load functions. If loading a backup fails, should we try to re-load it using these functions? Default = FALSE.}
}
\description{
Function to reload object(s) saved using saveFun() or saveImgFun().
The backend used is qs2::qs_savem under Windows and fastSave::save.lbzip2 under Linux.
}
\examples{
a <- "Hello world!"
proteoCraft::saveFun(a, "test.qs2")
rm(a)
proteoCraft::loadFun("test.qs2")
print(a)

}
