% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_ratio_av.R
\name{log_ratio_av}
\alias{log_ratio_av}
\title{log_ratio_av}
\usage{
log_ratio_av(x)
}
\arguments{
\item{x}{Log-transformed ratio values to be averaged.}
}
\description{
Averages log-transformed ratio values.
}
