% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.op.R
\name{logical.op}
\alias{logical.op}
\title{logical.op}
\usage{
logical.op(x, test, y, ties = TRUE)
}
\arguments{
\item{x}{Left-hand side term.}

\item{y}{Right-hand side term.}

\item{ties}{Logical: are ties accepted?}

\item{name}{Name of the operator.}
}
\description{
A function to call logical operators by name.
}
\examples{
logical.op("equals", x = 1, y = 2)

}
