% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Rat.R
\name{make_Rat}
\alias{make_Rat}
\title{make_Rat}
\usage{
make_Rat(
  myData = pep,
  Priority = "int",
  int.log = 10,
  rat.log = 2,
  refGroups,
  experiment.map,
  experiment.map_col = "Ref.Sample.Aggregate",
  int.root,
  rat.root,
  verbose = TRUE
)
}
\arguments{
\item{myData}{Input data.}

\item{Priority}{If set to "int", ratios calculated will reflect the real ratios of intensities. If set to "rat", intensities will be recalculated to reflect ratios.}

\item{int.log}{Input intensities log scale, default = 10}

\item{rat.log}{Input ratios log scale, default = 2}

\item{refGroups}{Aggregate defining ratio groups.}

\item{experiment.map}{Experiment map.}

\item{experiment.map_col}{Name of the sample column in the experiment map.}

\item{int.root}{Root of intensity column names.}

\item{rat.root}{Root of ratios column names.}

\item{verbose}{Logical, default = TRUE}
}
\value{
A list of two objects:
 - Ratios_to_Refs: map of ratios to references,
 - Data: data supplemented with new ratios and intensities columns
}
\description{
Internal function to output consistent ratios and intensity values.
}
