% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Rat2.R
\name{make_Rat2}
\alias{make_Rat2}
\title{make_Rat2}
\usage{
make_Rat2(
  myData = pep,
  contrasts = myContrasts,
  refGroups = RRG,
  int.log = 10L,
  rat.log = 2L,
  experiment.map,
  experiment.map_col = "Ref.Sample.Aggregate",
  int.root,
  rat.root
)
}
\arguments{
\item{myData}{Input data.}

\item{contrasts}{Contrasts data.frame for which to calculate ratios. It should contain columns of the form "[A-D]_samples" mapping samples to contrasts.}

\item{refGroups}{Factors aggregate defining blocks.}

\item{int.log}{Intensities log base, default = 10. Set to FALSE if input intensities are not log-transformed.}

\item{rat.log}{Ratios log base, default = 2}

\item{experiment.map}{Entire experiment map (unfiltered).}

\item{experiment.map_col}{Name of the sample column in the experiment map.}

\item{int.root}{Root of input intensity column names.}

\item{rat.root}{Root of output ratios column names.}
}
\value{
Data supplemented with new ratio columns. Unlike make_Rat, no intensities are output.
}
\description{
Internal function to output average log ratios.
This partially replaces the slowly deprecated make_Rat() function.
}
