% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMet_LCMS.R
\name{MatMet_LCMS}
\alias{MatMet_LCMS}
\title{MatMet_LCMS}
\usage{
MatMet_LCMS(
  ScanHdsMnLoc = "C:/ScanHeadsman-1.2.20200730",
  LocalRoot = "...Search_Folder/",
  ArchiveRoot = "...Archive/Acquired data/",
  RawFiles = rawFiles,
  Columns = "default",
  N.clust,
  N.reserved = 1,
  cl
)
}
\arguments{
\item{ScanHdsMnLoc}{ScanHeadsman path. Default = "C:/ScanHeadsman-1.2.20200730"; so far for us 1.2 works, 1.3 does not.}

\item{LocalRoot}{What is the root of the folder structure in the local folder. If the raw files are not local but were moved to an archive, this root will be replaced by that of the archive, see ArchiveRoot. Default = "D:/groups_temp/"}

\item{ArchiveRoot}{If the raw files are not local but were moved to an archive, where would that be? Default = "Q:/MS/Acquired data/"}

\item{RawFiles}{The raw files from which to attempt to extract the method.}

\item{Columns}{Path to a table of LC columns to pick from.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}
}
\value{
A list with two entries:
- "Text": the text for each LCMS method
- "Instruments": all LC and MS instrument models.
}
\description{
A function to write the LC-MS part of Materials & Methods, pulling information from the methods file.

Note to self:
!!! When editing function, always save with UTF-8 encoding!!!
}
\details{
Only tested on Bruker timsTOF HT with nanoElute 2 and Thermo Q-Exactive HT with Ultimate 3000 or Vanquish, so will probably fail on different machines from the same vendors, and obviously also on machines from other vendors.
}
\examples{
MatMet_Text_list <- MatMet_LCMS()

}
