% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMet_WetLab.R
\name{MatMet_WetLab}
\alias{MatMet_WetLab}
\title{MatMet_WetLab}
\usage{
MatMet_WetLab(
  File2Reload = "Materials and methods_WIP.docx",
  exp.map = Exp.map,
  exp.map.col = "MQ.Exp",
  Label = "LFQ",
  Columns = "default"
)
}
\arguments{
\item{File2Reload}{Optional file to reload the text from so you do not have to rewrite the whole thing. Default = "Materials and methods_WIP.docx".}

\item{exp.map}{The experiments map.}

\item{exp.map.col}{Name of the column in exp.map which contains biological sample names. Default = "MQ.Exp".}

\item{Label}{Were the samples labelled isobarically ("TMT", "TMTPRO" or "ITRAQ"), with "SILAC", or label-free ("LFQ")? This is only relevant for isobaric labels (otherwise ignored) because a) this affects the choice of iST kit, and b) the labelling procedure will then be reported in the samples processing.}

\item{Columns}{Path to a table of LC columns to pick from.}
}
\description{
A function to handle writing the wet-lab part of Materials & Methods, interactively asking relevant questions as it moves down the decision tree.
Very much a Rube Goldberg engine, which is fine. Just remember that it creates a WIP which should be checked/corrected before publication!

!!! When editing function, always save with UTF-8 encoding!!!
}
\examples{
if (LabelType == "LFQ") { MatMet_Text1 <- MatMet_WetLab() }
if (LabelType == "Isobaric") { MatMet_Text1 <- MatMet_WetLab(Label = IsobarLab) }

}
