% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ.load.R
\name{MQ.load}
\alias{MQ.load}
\title{MQ.load}
\usage{
MQ.load(
  ev = TRUE,
  ev.name = "ev",
  ev.file.name = "evidence.txt",
  pep = TRUE,
  pep.name = "pep",
  pep.file.name = "peptides.txt",
  prot = TRUE,
  prot.name = "PG",
  prot.file.name = "proteinGroups.txt",
  sub = FALSE,
  assign = TRUE,
  envir = .GlobalEnv,
  return = FALSE,
  check.names = FALSE
)
}
\arguments{
\item{ev}{Default = TRUE. If set to FALSE, the evidence file will not be processed.}

\item{ev.name}{The name of the evidence object that should be created in the above environment, if "assign" = TRUE.}

\item{ev.file.name}{The name of the evidence file that should be loaded.}

\item{pep}{Default = TRUE. If set to FALSE, the peptide file will not be processed.}

\item{pep.name}{The name of the peptides object that should be created in the above environment, if "assign" = TRUE.}

\item{pep.file.name}{The name of the peptides file that should be loaded.}

\item{prot}{Default = TRUE. If set to FALSE, the protein groups file will not be processed.}

\item{prot.name}{The name of the protein groups object that should be created in the above environment, if "assign" = TRUE.}

\item{prot.file.name}{The name of the protein groups file that should be loaded.}

\item{sub}{Are the files in a subfolder? Default = FALSE. If set to TRUE, will search in subfolders automatically, and will return an error if it finds more than one or no matches; can instead be set to any string specifying a subfolder.}

\item{assign}{Should we assign the files (defaut = TRUE) to a variable in the parent frame? Use arguments "ev.name", "pep.name" and "prot.name" to define the variable names.}

\item{return}{Should we return the files (in case the function is used in the form "object <- MQ.load(...)"); default = FALSE.}

\item{check.names}{logical (default = FALSE). If TRUE, then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names. If necessary they are adjusted (by make.names) so that they are, and also to ensure that there are no duplicates.}
}
\value{
Nothing is returned by default (the objects are assigned to the global environment). But if return = TRUE, then a list containing the "protein.groups", "peptides" and "evidences" data.frames (depending on whether they are to be processed or not) is returned.
}
\description{
A function to quickly import MaxQuant's main output datatables into the above environment.
}
\examples{
MQ.load()
# Result: the environment should now contain ev, pep and prot objects.

}
