% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ.summary.R
\name{MQ.summary}
\alias{MQ.summary}
\title{MQ.summary}
\usage{
MQ.summary(
  wd,
  ev,
  pg,
  filter = FALSE,
  mods = setNames(c("ox", "ac", "de", "gl", "ph"), c("Oxidation (M)",
    "Acetyl (Protein N-term)", "Deamidation (NQ)", "Gln->pyro-Glu", "Phospho (STY)")),
  raw.files,
  subfolder = "Summary plots",
  plot = TRUE,
  save = c("svg", "html"),
  sc = 60L,
  N.clust,
  N.reserved = 1L,
  cl,
  MQtxt = inDirs[which(SearchSoft == "MAXQUANT")]
)
}
\arguments{
\item{wd}{Working directory. Automatically set to local if left empty.}

\item{ev}{The evidence/PSMs file. Will try to load from the local directory if value is missing.}

\item{pg}{The protein groups file. Will try to load from the local directory if value is missing.}

\item{filter}{Should we filter the files for contaminants, reverse hits, NA values, etc... Default = FALSE}

\item{mods}{(Ideally named) vector of 2-letter modification codes (as used by older MaxQuant versions for Modified Sequences) of interest. The default is the default MQ PTMs list which we use in most projects.}

\item{raw.files}{Vector of full raw file paths, sorted in the order you will want to use.}

\item{plot}{Should we create plot?}

\item{save}{Set this to a vector of acceptable file extensions to save the graph to the corresponding file format, or to FALSE if you do not want to save it. Default = c("svg", "html")}

\item{sc}{Scale, the max number of raw files to be plotted together in one plot. Default = 60.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{MQtxt}{For MaxQuant only: the directory in which to search for MaxQuant evidence.txt, msms.txt and msmsScans.txt files.}
}
\description{
A function to summarize the data from an MS experiment. Originally limited to MaxQuant searches, hence the name.
Now also works partially for the output of DIA-NN if converted to MaxQuant-like format.
}
\examples{
## To load data from local directory and filter it:
MQ.summary(filter = TRUE)
## To summarize processed data, already loaded in the environment, and look at phospho-peptides
MQ.summary(ev = ev, pg = PG, mods = "ph")

}
