% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(
  ...,
  plotlist = NULL,
  cols = 1,
  layout = NULL,
  save = FALSE,
  filename = NULL
)
}
\arguments{
\item{plotlist}{The plots to plot. Duh!}

\item{cols}{Number of columns. The number of rows will be inferred from this.}

\item{layout}{A layout matrix. Usually assumed to be null so created by the function.}

\item{save}{FALSE by default. Set to format you want pictures saved as.}

\item{filename}{Required only if save = TRUE. Extensions will be ignored as they are provided by the save argument.}
}
\description{
I did not actually create this function but nicked it from the R cookbook, so credit goes to them:
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)
Currently unused, thus candidate for deletion.
}
