% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.mod.R
\name{norm.mod}
\alias{norm.mod}
\title{norm.mod}
\usage{
norm.mod(values, bins = 100, mu = FALSE, adjust.bins = TRUE, use.frac = 0.9)
}
\arguments{
\item{values}{The vector of normally distributed data}

\item{bins}{How many bins should we make? Default = 100.}

\item{mu}{The mu parameter of the normal distribution. Usually I will be using this function on data where I want mu to be 0 (assumed if input mu is FALSE). If TRUE, the estimate will return a mu value.}

\item{adjust.bins}{TRUE by default. If set, will adjust bin number so there is at least a ratio of points to bins of 5, and at least 3 bins; failing that it will throw an error.}

\item{use.frac}{What fraction of the data to use (from the center)? It will exclude outliers to avoid spreading the bins too far to both sides, so we get good sampling in the centre. Default is 0.9.}
}
\description{
Create a normal distribution model for normally distributed data.
The function assumes the data is normal, and will thus NOT check for normality by itself.
}
\examples{
norm.mod(values = x)

}
