% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.mod2.R
\name{norm.mod2}
\alias{norm.mod2}
\title{norm.mod2}
\usage{
norm.mod2(
  values,
  shapiro.threshold = 0.05,
  print.Density = FALSE,
  print.QQ = FALSE
)
}
\arguments{
\item{values}{The vector of normally distributed data}

\item{shapiro.threshold}{Which threshold can we accept for the Shapiro-Wilk normality test's P-value? Default = 0.05}

\item{print.Density}{Should we print the density plot? Default = FALSE.}

\item{print.QQ}{Should we print the Q-Q plot? Default = FALSE.}
}
\description{
Create a normal distribution model for normally distributed data.
}
\examples{
norm.mod2(values = x)

}
