% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.thresh.R
\name{norm.thresh}
\alias{norm.thresh}
\title{norm.thresh}
\usage{
norm.thresh(values, ref = NULL, p = 0.01, shapiro.threshold = 0.05)
}
\arguments{
\item{values}{The vector of normally distributed values for which we want a threshold.}

\item{ref}{The reference. If left as NULL, proportions are from the values vector. If provided as a vector of values, the function aims to find thresholds including a given proportion (based on modelled normal distributions) p of reference to values.}

\item{p}{The relevant proportion to include (either absolute, or contamination by a reference).}

\item{shapiro.threshold}{Which threshold can we accept for the Shapiro-Wilk normality test's P-value? Default = 0.05}
}
\description{
Find thresholds to include only a given proportion of data.
}
\examples{
norm.thresh(values = test$"log2.Ratio.Exp1 Astrocyte ApoE4/4",
              ref = test$"log2.Ratio.Exp1 Astrocyte ApoE3/3")

}
