% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwd.R
\name{openwd}
\alias{openwd}
\title{openwd}
\usage{
openwd(dir)
}
\arguments{
\item{dir}{Directory to open. If missing, defaults to current work directory.}
}
\description{
A function to open your a directory (work directory by default). Only works on PCs.
}
\examples{
openwd() # Open current work directory
openwd("C:/MaxQuant/")

}
