% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpMS_to_MQ.R
\name{OpMS_to_MQ}
\alias{OpMS_to_MQ}
\title{OpMS_to_MQ}
\usage{
OpMS_to_MQ(
  OpMS,
  Fixed.mods = c(paste0("tmt", c("", paste0(0:11, "plex"), paste0("pro", 0:16, "plex"))),
    paste0("itraq", c("", paste0(0:8, "plex"))), "carbamidomethyl", "nethylmaleimide"),
  Contaminants = "MQ",
  MQ_loc = "C:/MaxQuant",
  FixUniProtIDs = TRUE,
  Isobaric = FALSE,
  Label,
  MaxRank = 0
)
}
\arguments{
\item{OpMS}{The name of a results csv file from OpenMS.}

\item{Fixed.mods}{Fixed modifications. Will not be reported in the results.}

\item{Contaminants}{Database of common contaminant proteins used for the search. Set either to "MQ" (default; in this case the next argument is used), or directly to a valid fasta database.}

\item{MQ_loc}{Location of your MaxQuant installation. Default = "C:/MaxQuant". It then navigates down to the normal MaxQuant "contaminants.fasta" database. If several MaxQuant versions are detected in subfolders (format expected: "MaxQuant_W.X.Y.Z"), it will determine the latest and used it by default.}

\item{FixUniProtIDs}{If set to TRUE, UniProtKB IDs of the form "xy|ACCESS|NAME_SPECIE" will be changed into the shorter "ACCESS" (for accession) format.}

\item{Isobaric}{Are there isobaric labelling reporter intensity columns in the table? Default = FALSE}

\item{Label}{This parameter is required if Isobaric is set to TRUE. Accepted values are: "TMT0plex", "TMTpro0plex", "TMT2plex", "TMT6to11plex", "TMTpro16plex", "iodoTMT6plex", "iodoTMT6plex", "aminoxyTMT0plex", "aminoxyTMT6plex", "iTRAQ4plex" or "iTRAQ8plex"}

\item{MaxRank}{Will only accept search engine ranks lower or equal to the value of this parameter (default = 1)}
}
\description{
Very old/deprecated, but could be the basis of an updated function should we start working with OpenMS again.
Process OpenMS csv output file into a valid PSM file
Not yet tested for:
- Labelled samples.
- Cases where a single amino-acid has been assigned multiple PTMs.
}
\examples{
temp <- OpMS_to_MQ("Results.csv")
ev <- temp$Evidences
Modifs <- temp$PTMs

}
