% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param.load.R
\name{Param.load}
\alias{Param.load}
\title{Param.load}
\usage{
Param.load(file = "Parameters.csv", filter.deprecated = TRUE, WD_detect = TRUE)
}
\arguments{
\item{file}{The parameters file. Must be a txt or csv file.}

\item{filter.deprecated}{Should we remove deprecated parameters? Default = TRUE}

\item{WD_detect}{If set to TRUE (default), will either set work directory to that contained in the parameter file (if there), or will set the parameter value "WD" to the current one.}
}
\description{
A function to quickly import analysis Parameters into the environment.
}
\examples{
Param <- Param.load()

}
