% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_filter.R
\name{Param_filter}
\alias{Param_filter}
\title{Param_filter}
\usage{
Param_filter(param, filter)
}
\arguments{
\item{param}{The parameter to filter.}

\item{filter}{Which factor(s) to remove.}
}
\description{
A simple function to remove some factors from semicolon-separated factor aggregate parameters.
}
\examples{
Param$Ratios.Groups.Ref.Aggregate.Level
# [1] "Exp;Con;Rep"
Param_filter(Param$Ratios.Groups.Ref.Aggregate.Level, "Rep")
# [1] "Exp;Con"

}
