% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_to_MQ.R
\name{PD_to_MQ}
\alias{PD_to_MQ}
\title{PD_to_MQ}
\usage{
PD_to_MQ(
  PD,
  Fixed.mods = c(paste0("TMT", c("", paste0(0:11, "plex"), paste0("pro", 0:16, "plex"))),
    paste0("iTRAQ", c("", paste0(0:8, "plex"))), "Carbamidomethyl", "N-Ethylmaleimide"),
  Abs.quant.col = "Area",
  Isobaric = FALSE,
  Label,
  filter = FALSE,
  MaxRank = 1
)
}
\arguments{
\item{PD}{A PSMs table from Proteome Discover imported as a data frame.}

\item{Fixed.mods}{Fixed modifications which will not be reported in the results (PD reports isobaric tags and fixed modifications, MQ ignores them). Default covers "TMT0plex" to "TMT11plex", "iTRAQ0plex" to "iTRAQ8plex", "Carbamidomethyl" and "N-Ethylmaleimide".}

\item{Abs.quant.col}{Default = "Area". If not provided, and if reporter intensities are available, we will sum those instead.}

\item{Isobaric}{Are there isobaric labelling reporter intensity columns in the table? Default = FALSE}

\item{Label}{This parameter is required if Isobaric is set to TRUE. Accepted values are: "TMT0plex", "TMTpro0plex", "TMT2plex", "TMT6to11plex", "TMTpro16plex", "iodoTMT6plex", "iodoTMT6plex", "aminoxyTMT0plex", "aminoxyTMT6plex", "iTRAQ4plex" or "iTRAQ8plex"}

\item{filter}{If set to TRUE, this will remove all PSMs flagged as "Rejected". Default = FALSE}

\item{MaxRank}{Will only accept search engine ranks lower or equal to the value of this parameter (default = 1)}
}
\description{
Converts a Proteome Discoverer PSMs table to a MaxQuant evidence.txt-like table.
As with the other functions of this type, the idea is not to get perfect conversion, but close enough that the data can be fed into this package's analysis scripts.

Has not been updated for a long time, and is probably in need of a refresh!

The output is a list of two items:
- an evidence table similar to the one MaxQuant creates ("Evidences")
- a modifications table ("PTMs")
}
\examples{
temp <- PD_to_MQ(PD)
ev <- temp$Evidences
Modifs <- temp$PTMs

}
