% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepHtmp.R
\name{pepHtmp}
\alias{pepHtmp}
\title{pepHtmp}
\usage{
pepHtmp(
  intProt = prot.list_pep,
  Pep = pep,
  ref = pep.ref[length(pep.ref)],
  dstDir,
  ttlRoot = "Peptides log2 heatmap",
  ttlStem = "",
  DB = db,
  Experiment = Exp,
  RSA = Ref.Sample.Aggregate,
  is.log = FALSE,
  cl,
  N.clust,
  N.reserved = 1L
)
}
\arguments{
\item{intProt}{Vector of protein accessions of interest.}

\item{Pep}{Peptides data.frame, default = pep}

\item{ref}{Roof of intensity column names, default = pep.ref[length(pep.ref)]}

\item{dstDir}{Destination directory.}

\item{ttlRoot}{Prefix of the plot title, default = "Peptides log2 heatmap"}

\item{ttlStem}{A suffix to add to the title.}

\item{DB}{Parsed fasta data.base, default = db}

\item{Experiment}{= Exp}

\item{RSA}{Ref.Sample.Aggregate object, default = Ref.Sample.Aggregate}

\item{is.log}{Logical, is the data log-transformed? (default = FALSE)}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}
}
\value{
This function does not return anything.
}
\description{
Peptides heatmap function - useful for checking normalisations.
}
