% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pepper_ProcessData.R
\name{Pepper_ProcessData}
\alias{Pepper_ProcessData}
\title{Pepper_ProcessData}
\usage{
Pepper_ProcessData(
  Ev,
  Modifs,
  experiments.map,
  Ev_Sample.col,
  experiments.map_Sample.col,
  path = "",
  SearchSoft = SearchSoft,
  MQFold,
  intCol = "Intensity",
  filter = FALSE
)
}
\arguments{
\item{Ev}{The evidences (PSMs) data frame.}

\item{Modifs}{Modifications table as output by DIANN_to_MQ, FP_to_MQ, PD_to_MQ... Should ideally contain a "Mass shift" column, but for MQ workflows this will be usually missing: two additional arguments are provided to deal with those cases: SearchSoft and MQFold.}

\item{experiments.map}{The experiments map.}

\item{Ev_Sample.col}{Name of the sample column in Ev}

\item{experiments.map_Sample.col}{Name of the sample column in experiments.map}

\item{path}{The path to which to save the file.}

\item{SearchSoft}{Search software used. Important because Modifs may miss the "Mass shift" column for data originating from some software (MaxQuant). If it is missing, and software is MaxQuant, then providing the next argument (MQFold) can allow parsing directly modifications and calculating the associated mass shift.}

\item{MQFold}{Main MaxQuant folder, containing /bin/conf subfolder with modification xml(s). Only used if SearchSoft == "MaxQuant"}

\item{intCol}{Which intensity column to use? Default = "Intensity"}

\item{filter}{If TRUE, the data is filtered to create a training set. If FALSE (default), the data isn't filtered (used for preparing data to which a pre-trained model will be transferred).}
}
\description{
A function to take an ev data.frame and convert it to a Pepper-compatible table. This can then be one-hot encoded for the purpose
of training a model on it, or a pre-trained model can be applied to it.


Note:
An issue may be silent fixed modifications (TMT, cysteine alkylations) in MaxQuant. The final model should thus specify
what type of data it is suitable for.
}
