% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepPlotFun.R
\name{pepPlotFun}
\alias{pepPlotFun}
\title{pepPlotFun}
\usage{
pepPlotFun(
  df1,
  df2,
  ttl,
  dstDir,
  save = TRUE,
  xpMap = Exp.map,
  VPAL = Volcano.plots.Aggregate.Level
)
}
\arguments{
\item{df1}{Quantitative data prior to normalisation.}

\item{df2}{Normalized quantitative data.}

\item{ttl}{Title.}

\item{dstDir}{Destination directory.}

\item{save}{Logical (default = TRUE), should we save the plot. Currently, the format is svg.}

\item{xpMap}{Experiment map, default = Exp.map}

\item{VPAL}{Volcano.plots.Aggregate.Level object, default = Volcano.plots.Aggregate.Level}
}
\value{
This function does not return anything.
}
\description{
Useful function for checking before/after peptides normalisation.
}
