% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PG_assemble.R
\name{PG_assemble}
\alias{PG_assemble}
\title{PG_assemble}
\usage{
PG_assemble(
  Pep,
  Peptide.IDs = "id",
  Proteins.col = "Proteins",
  DB,
  Evidence.IDs = "Evidence IDs",
  Ev,
  N.clust,
  N.reserved = 1,
  Custom_PGs = NA,
  ContCol = "Potential contaminant",
  Npep = 2,
  cl
)
}
\arguments{
\item{Pep}{The peptides sequences data frame. Should contain "Modified.sequence", "Sequence" and "PEP" columns, as well as IDs, protein and evidence IDs columns (see next two arguments).}

\item{Peptide.IDs}{Name of the column with peptide IDs. Default = "id"}

\item{Proteins.col}{Name of the column with matched protein accessions. Default = "Proteins"}

\item{DB}{A search database parsed as a table by the Format.DB function. Will also act as a filter: only protein accessions present in its "Protein ID" column will be part of the assembly!!!}

\item{Evidence.IDs}{Name of the column with evidence IDs. Default = "Evidence IDs"}

\item{Ev}{The evidences data frame. Used for a few annotations columns. Not required.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{Custom_PGs}{Data frame of protein groups known a priori (default = NA). One column required, called "Leading protein IDs" with semicolon-separated IDs; optional 2nd column: a numeric/integer called "Priority level". Any peptide mapping to these groups will be assigned to them to the exclusion of other protein groups. Use the priority column to resolve ambiguities (1 = highest priority). NB: Use with caution, only for known, priority proteins (e.g. artificial constructs).}

\item{ContCol}{Name of column in the database specifying whether a protein is a contaminant ("+", otherwise ""). Default = "Potential contaminant"}

\item{Npep}{Minimum number of razor or unique peptides to tag a protein group as a true discovery (others will still be reported). Default = 2}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}
}
\value{
A list with 3/4 slots:
 - "Protein.groups": data.frame of inferred protein groups
 - "Peptides": input peptides data.frame with a few extra columns
 - "Database": input database data.frame with a few extra columns
 - "Evidences" (only if Ev was provided): input PSMs data.frame with a few extra columns
}
\description{
Schedule for slow retirement: the code for this function is now in a source because of parallelisation issues.
For now this is still kept for older workflows.
In cases of bugs check first for discrepancy between the more up-to-date source and the function.

 Note:
 Pep expects a wide (peptides), rather than long (PSMs/evidences), table! Ev is only used for mapping PSMs, not to assemble protein groups.
 We haven't explored the potential implications for the function's logic of supplying to Pep a long table.
 But when this was accidentally tested on a large-ish dataset, a powerful PC quickly froze and did not manage to resolve the task overnight.
 Just do not do it!
}
\examples{
temp <- PG_assemble(Pep = pep, DB = db)

}
