% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phos_QC.R
\name{phos_QC}
\alias{phos_QC}
\title{phos_QC}
\usage{
phos_QC(
  PhPep,
  min_P = 0.75,
  min_delta_sc = 5,
  max_PEP = 0.01,
  ModSeq_col = "Modified sequence",
  P_col = "Phospho (STY) Probabilities",
  ScD_col = "Phospho (STY) Score Diffs",
  PEP_col = "PEP"
)
}
\arguments{
\item{PhPep}{The phospho-peptides table.}

\item{min_P}{The minimum site probability acceptable. Default = 0.75}

\item{min_delta_sc}{The minimum delta score acceptable. Default = 5}

\item{max_PEP}{The maximum Posterior Error Probability acceptable. Default = 0.01}

\item{ModSeq_col}{The name of the modified sequence column. Default = "Modified sequence"}

\item{P_col}{The name of the site probabilities column. Default = "Phospho (STY) Probabilities"}

\item{ScD_col}{The name of the score differences column. Default = "Phospho (STY) Score Diffs"}

\item{PEP_col}{The name of the Posterior Error Probability column. Default = "PEP"}
}
\description{
A function to create a column saying whether a phospho-peptide is high quality or not.
Useful for creating a table of, e.g. sequence with amino-acid modifications, site probabilities, etc...
}
\examples{
phospep$High_Quality <- phos_QC(phospep)

}
