% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEval.R
\name{plotEval}
\alias{plotEval}
\title{plotEval}
\usage{
plotEval(plot, ggplotify = FALSE)
}
\arguments{
\item{plot}{The ggplot to evaluate.}

\item{ggplotify}{Alternate (classic) method using ggplotify, producing a more complete (larger) evaluated ggplot.}
}
\value{
Evaluated ggplot.
}
\description{
A simple wrapup for evaluation a ggplot, allowing to save it in a parallel manner.
By default this now outputs a gtable, but may also output a full ggplot (if ggplotify is TRUE).
}
