% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplot.R
\name{poplot}
\alias{poplot}
\title{poplot}
\usage{
poplot(plot, height = 10, width = 10, new.window = TRUE, max.gr.dev = 62)
}
\arguments{
\item{plot}{ggplot to print.}

\item{height}{Height of the popup. Default = 10.}

\item{width}{Width of the popup. Default = 15.}

\item{new.window}{TRUE by default, set to FALSE to not plot in a popup window.}

\item{max.gr.dev}{62 by default, the maximum number of graphical devices which the system will allow to be open simultaneously. I could not find the information anywhere but checked on 3 different PCs and the result was always 62.}
}
\description{
A wrapup for printing a ggplot in a popup window (or not: can also just do in local window). Should work on all devices (some strange people like Macs.)
}
