% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prot.Quant2.R
\name{Prot.Quant2}
\alias{Prot.Quant2}
\title{Prot.Quant2}
\usage{
Prot.Quant2(
  Prot,
  Peptide.IDs = "Peptide IDs",
  Pep,
  id = "id",
  Summary.method = "median",
  Summary.weights,
  Priority = "Intensities",
  Skip.ratios = FALSE,
  experiments.map,
  ref.groups,
  ratio.groups,
  param,
  aggregate.map = Aggregate.map,
  aggregate.list = Aggregate.list,
  aggregates = Aggregates,
  Pep.Intens.root,
  Pep.Ratios.root,
  log.Pep.Intens = FALSE,
  log.Pep.Ratios = 2,
  Prot.LFQ.to.log = TRUE,
  Prot.Ratios.to.log = TRUE,
  Mods,
  Mods.to.Exclude,
  Mod.Nms = "Modified sequence",
  Discard.unmod = TRUE,
  Prot.Seq = "Sequence (1st accession)",
  Min.N = 2
)
}
\arguments{
\item{Prot}{The protein/protein groups file.}

\item{Peptide.IDs}{The name of the Protein/Protein groups file's peptide IDs column. Default = "Peptide.IDs"}

\item{Pep}{The peptides file.}

\item{id}{The name of the Peptides file IDs column. Default = "id"}

\item{Summary.method}{The summary method used for ratios (sum or weighted sum is used for intensities depending on whether weights are provided). One of "mean", "median", or "weighted.mean".}

\item{Summary.weights}{If a "weighted.mean" summary method is chosen, then a vector of weights must be provided. This should be the name of a column of the peptides file.}

\item{Priority}{One of "Ratios" or "Intensities" (default). Some flexibility in spelling is allowed. You want to prioritize ratios for SILAC because in this case MaxQuant measures peptides ratios more precisely than intensities. Otherwise, you want to prioritize Intensities and re-calculate ratios from them.}

\item{Skip.ratios}{Default = FALSE. If TRUE, ratios will not be calculated.}

\item{experiments.map}{The experiments map.}

\item{ref.groups}{Defines which samples are paired to which references. May alternatively (preferred solution) be provided indirectly through the param argument.}

\item{ratio.groups}{Defines groups within which ratios are calculated. May alternatively (preferred solution) be provided indirectly through the param argument.}

\item{param}{The experiment's parameters file. If provided, the ref.groups argument is not required.}

\item{aggregate.map}{The aggregate map. Default = Aggregate.map}

\item{aggregate.list}{The named list of aggregates. Default = Aggregate.list}

\item{aggregates}{The aggregates themselves. Default = Aggregates}

\item{Pep.Intens.root}{Root of the peptides intensity column(s) names}

\item{Pep.Ratios.root}{Root of the peptides ratios column(s) names}

\item{log.Pep.Intens}{Set to 0 or FALSE if input peptide intensities are linear scale. If the data is already log scale, set to the relevant scale's base. Default = FALSE}

\item{log.Pep.Ratios}{Set to 0 or FALSE if input peptide ratios are linear scale. If the data is already log scale, set to the relevant scale's base. Default = 2}

\item{Prot.LFQ.to.log}{Should the output protein LFQ values be log-scale or not? Can ve set to the log base desired. If set to TRUE, this will be the same base as the input intensities log scale, or 10 by default.}

\item{Prot.Ratios.to.log}{Should the output protein ratios be log-scale or not? Can ve set to the log base desired. If set to TRUE, this will be the same base as the input ratios log scale, or its default, 2}

\item{Mods}{Which modifications (2 lowercase letters PTM code) should be included? If set to FALSE, will not filter any modifications.}

\item{Mods.to.Exclude}{Alternative/complementary to Mods. Which modifications should be excluded? (use argument "Discard.unmod" to discard unmodified counterpart peptides.) A data.frame with columns "Mark" (code) and "Where", see argument "Discard.unmod".}

\item{Mod.Nms}{Default = "Modified sequence". The name of the column containing the modified sequence in the peptides table. Can be set to a non-modified sequence if "Mods" = FALSE and "Mods.to.Exclude" is empty.}

\item{Discard.unmod}{Default = TRUE. Should we discard those unmodified peptides whose primary sequence is the same as that of some modified peptides we will not use? If set to 2, will use the "Where" column in "Mods.to.Exclude" to identify (and exclude) peptides which could be modified even if the modified form was not identified. Requires knowledge of protein sequence ("Prot.Seq" argument)! Be careful! This will likely massively reduce the number of peptides available for quantitation!}

\item{Prot.Seq}{Default = "Sequence (1st accession)", used if "Discard.unmod" is set to 2}

\item{Min.N}{How many peptides should at least be present? Should be at the very least 1 (default = 2).}
}
\description{
A function to calculate estimated protein group Expression (Absolute Quant) and Ratios (optional) from individual peptide values.
The input is assumed to be normalized.

A note on ref.groups and ratio.groups:
These parameters are the method I have used to define if replicates are paired or not.
For an experiment with one master Experiment, several Conditions (incl. one reference) and Replicates:
- in a paired (= "nested") setup, you would set ref.groups to "ExpRep" and ratio.groups to "Exp"
- in an unpaired setup, you would set ref.groups to "Exp" to and ratio.groups to "Exp"
Essentially, ref.groups are the groups within which ratios are calculated to all available references,
while ratio.groups are groups within which reference-to-reference values are calculated:
- between individual references in the paired setup
- from individual references to the average reference in the unpaired setup
}
\examples{
temp <- Prot.Quant2(Prot = PG, Peptide.IDs = Pep4Quant, Pep = pep, id = "New Peptide ID",
                   experiments.map = Exp.map,
                   ref.groups = Ratios.Ref.Groups,
                   Pep.Intens.root = pep.ref, Pep.Ratios.root = pep.ratios.ref,
                   log.Pep.Intens = FALSE, log.Pep.Ratios = 2,
                   Prot.LFQ.to.log = TRUE, Prot.Ratios.to.log = TRUE,
                   Mods = Mod4Quant, Mod.Nms = "Modified sequence",
                   Min.N = 2)

}
