% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prot.Ratios.R
\name{Prot.Ratios}
\alias{Prot.Ratios}
\title{Prot.Ratios}
\usage{
Prot.Ratios(
  Prot,
  Peptide.IDs = "Peptide IDs",
  Pep,
  id = "id",
  Pep.Ratios.Nms,
  Pep.Ratios.to.log = TRUE,
  Prot.Ratios.to.log = TRUE,
  Mods,
  Mod.Nms = "Modified sequence",
  Min.Pep.Nb = 2,
  SD = TRUE,
  Pvalue = TRUE,
  In.Norm = FALSE,
  Out.Norm = TRUE
)
}
\arguments{
\item{Prot}{The protein/protein groups file.}

\item{Peptide.IDs}{The name of the Protein/Protein groups file's peptide IDs column. Default = "Peptide IDs"}

\item{Pep}{The peptides file.}

\item{id}{The name of the Peptides file IDs column. Default = "id"}

\item{Pep.Ratios.Nms}{Name of the peptides ratios column(s)}

\item{Pep.Ratios.to.log}{Should the input peptide ratios be converted to log scale (log2)? If they aren't, then they must, as the function will assume they are log scale. Default = TRUE}

\item{Prot.Ratios.to.log}{Should the output protein ratios be log-scale or not? Default = TRUE}

\item{Mods}{Which modifications should be included? If set to FALSE, will not filter any modifications.}

\item{Mod.Nms}{Default = "Modified sequence". The name of the column containing the Modified sequence in the peptides file. Can be set to a non-modified sequence if Mods = FALSE}

\item{Min.Pep.Nb}{How many peptides should at least be present? Should be at the very least 1, but really we don't advise going below 2 (set do default).}

\item{SD}{Should the Output also include the standard deviation? Default = TRUE}

\item{Pvalue}{Should the Output also include the P-value? Default = TRUE}

\item{In.Norm}{Should the Input be re-normalised? Default = FALSE}

\item{Out.Norm}{Should the Output be re-normalised? Default = TRUE}
}
\description{
A function to calculate protein SILAC Ratios from individual peptides.
NB: Use with caution, this is an old function which can be considered deprecated.
Anyway for SILAC ratios the safest way is to accept MaxQuant ratios as more accurate than the ratio of intensity values.
Candidate for deletion!
}
\examples{
test <- Prot.Ratios(Prot = prot, Peptide.IDs = "New.Peptide.IDs", Pep = pep, id = "New.Peptide.ID",
                    Pep.Ratios.Nms = paste0("log2.", sapply(Exp, function(x) {paste0("norm.Ratio.", x, ".", unique(unlist(strsplit(Ratios.map$Ratios, split = ";"))))})),
                    Pep.Ratios.to.log = FALSE, Prot.Ratios.to.log = TRUE,
                    Mods = Mod4Quant, Mod.Nms = "Modified.sequence", Min.Pep.Nb = 2, SD = TRUE, Pvalue = TRUE, In.Norm = TRUE, Out.Norm = TRUE)
colnames(test) <- gsub("Protein_", "", colnames(test))
prot[, colnames(test)] <- test

}
