% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtFormula.R
\name{ProtFormula}
\alias{ProtFormula}
\title{ProtFormula}
\usage{
ProtFormula(Seq, collapse = TRUE, reduced = TRUE, alkylated = "IAM")
}
\arguments{
\item{Seq}{Protein single letter code sequence.}

\item{collapse}{If TRUE, collapse as single character chain, otherwise provided as table of individual element counts.}

\item{reduced}{If TRUE, then disulfide bonds are assumed to be reduced. Note: We are assuming that all cysteines are parts of disulfide bonds... but we do not know that! One solution would be to provide a new way to map known disulfide bonds to the sequence. For now the function cannot handle this case other than by assuming that all cysteines are oxidized.}

\item{alkylated}{If one of "IAM", "NEM", "CAM", "MMTS" or "iST_NHS" (alkylating agent used in the iST-NHS kit), then the appropriate modification will be applied on cysteines.}
}
\description{
Calculate the chemical formula of a protein.
}
\examples{
Form <- ProtFormula(Seq, collapse = FALSE, reduced = FALSE)
Form <- ProtFormula(Seq, collapse = FALSE, alkylated = FALSE)
Form <- ProtFormula(Seq, collapse = FALSE)

}
