% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtMatch2.R
\name{ProtMatch2}
\alias{ProtMatch2}
\title{ProtMatch2}
\usage{
ProtMatch2(
  Seq,
  DB,
  Cut = c("K_", "R_"),
  N.clust = get("N.clust"),
  N.reserved = 1,
  cl
)
}
\arguments{
\item{Seq}{Character vector of peptide sequences.}

\item{DB}{The formated protein database, with proteins IDs and sequences.}

\item{Cut}{For digestion. The character vector of amino acid(s) after which to cut. For each, provide one letter followed or preceded by underscore indicating whether the cut is C- or N-terminal, respectively (usually C-terminal). Default (for Trypsin) = c("K_", "R_")}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1. Only used if missed > 0}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one. Only used if missed > 0}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}
}
\description{
Once again, the function has been radically re-written, though without renaming it this time!
The indexing from the previous version is gone, replaced by better parallelisation and data.tables use, yielding a small improvement.
A small, rare bug is also corrected in the process.
}
