% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustNorm.R
\name{robustNorm}
\alias{robustNorm}
\title{robustNorm}
\usage{
robustNorm(
  df,
  exprs.col,
  loss = c("l2", "huber", "tukey", "cauchy"),
  k = NULL,
  max_iter = 50L,
  tol = 1e-06
)
}
\arguments{
\item{df}{Data frame containing expression values.}

\item{exprs.col}{Name of the columns containing expression data. THIS SHOULD BE LOG-TRANSFORMED!}

\item{loss}{Loss function. Options are "l2" (= sum of squared errors), "huber", "tukey" or "cauchy".}

\item{k}{Tuning constants for the method, allowing to override the defaults.}

\item{max_iter}{Integer, default = 50L.}

\item{tol}{Tolerance, default = 1e-6}
}
\value{
A list:
 - data = data.frame with normalized values
 - sample_offset = optimized offsets
 - row_mean = row_mean,
 - weights
 - iterations = last iteration before convergence 
 - loss = loss function used
 - sigma = scaled Median Absolute Deviation of the residuals
}
\description{
This is meant as an alternative to AdvNorm.IL() for applying a more robust method to aligning the samples. AdvNorm.IL is currently faster though.
I have only tested, not written this function. ChatGPT and Perplexity did, in a merry-go-round, where I bounced the code from one to the other.
Yes.
We are there.
...
The idea for a more robust loss function was floated to my by chatGPT as I was discussing normalization with 'it'.
And because I was too busy with other stuff and unfamiliar with the robust loss function landscape, I asked it to make it for me.
}
\examples{
normData <- robustNorm(pep, paste0("log10(Int.) - ", mySamples))

}
