% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFun.R
\name{saveFun}
\alias{saveFun}
\title{saveFun}
\usage{
saveFun(x, file)
}
\arguments{
\item{x}{The object to save.}

\item{file}{The path/file to save to.}
}
\description{
Function to save objects quickly.
The backend used is qs2::qs_savem under Windows and fastSave::save.lbzip2 under Linux.
}
\examples{
a <- "Hello world!"
proteoCraft::saveFun(a, "test.qs2")
rm(a)
proteoCraft::loadFun("test.qs2")
print(a)

}
