% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveImgFun.R
\name{saveImgFun}
\alias{saveImgFun}
\title{saveImgFun}
\usage{
saveImgFun(file, N.clust, N.reserved = 1L)
}
\arguments{
\item{file}{The path/file to save to.}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1. This is probably not a good idea for large sessions however!}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}
}
\description{
Function to save all objects in the environment quickly.
The backend used is qs2::qs_savem under Windows and fastSave::save.lbzip2 under Linux.
}
\examples{
saveImgFun("Backup.qs2")
loadFun("Backup.qs2")

}
