% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_ions.R
\name{seq_ions}
\alias{seq_ions}
\title{seq_ions}
\usage{
seq_ions(
  Seq,
  Z,
  types = c("b", "y"),
  C_alkylation = "NEM",
  C_mass_shift = NULL,
  Mods,
  plot = FALSE
)
}
\arguments{
\item{Seq}{The peptide sequence. Modifications are now supported.}

\item{Z}{The parent peptide's charge.}

\item{types}{The type of sequence ions to compute. Available are "a", "b", "c", "x", "y", "z".}

\item{C_alkylation}{One of "NEM" (default), "IAA" or "custom". For the latter case, a mass shift should be provided.}

\item{C_mass_shift}{Only required if "C_alkylation" is "Custom", other wise will be ignored.}

\item{Mods}{A data.frame with two columns: "Code" should be the 2-letter code of the modifications as reported by MaxQuant in the "Modified.sequence" column. "Shift" should be the associated expected mass shift.}

\item{plot}{Should we print a plot of expected peak positions? FALSE by default.}
}
\description{
This function is not used, more of an unfinished exercise.
It should calculate the selected sequence ions for the peptide sequence, at all allowed charges.
It assumes a monoisotopic precursor.
This function has the following shortcomings - at least:
- It cannot accept TMT/iTRAQ modifications, or SILAC mass shifts.
- It cannot yet calculate internal cleavage, neutral loss or immonium fragments.
- It cannot predict fragment relative intensities... but that's kinda expected: who can? EDIT: well ML can now.
}
\examples{
Result <- seq_ions(Seq = "GLSDGEWQQVLNVWGK", Z = 3, types = c("b", "y"), plot = TRUE)

}
