% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCheckInput.R
\name{shinyCheckInput}
\alias{shinyCheckInput}
\title{shinyCheckInput}
\usage{
shinyCheckInput(startVal, root = "check")
}
\arguments{
\item{startVal}{Starting values of each checkbox.}

\item{root}{Root of the created input names. Names generated will be of the form "root___i", where i is the 1:n row number.}
}
\description{
Convenience wrapper for shiny::checkboxInput for use in DT data table columns.
}
\examples{
# Use within a shiny server as below:

server <- function(input, output, session) {

  Table$CheckBox <- shinyCheckInput(Table$UseThisSample)

}

}
