% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCleanup.R
\name{shinyCleanup}
\alias{shinyCleanup}
\title{shinyCleanup}
\usage{
shinyCleanup(max_iter = 100L)
}
\arguments{
\item{max_iter}{Max number of iterations, default = 100.}
}
\value{
This function does not return anything.
}
\description{
A function to clean up after a shiny app finishes running, so that the next one is not affected by leftover from the previous.
I do not have time for designing this type of function, so must admit that this is 100% from chatGPT.
}
