% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlusFD.R
\name{shinyPlusFD}
\alias{shinyPlusFD}
\title{shinyPlusFD}
\usage{
shinyPlusFD(root, l, buttonLabel = "+", stem = "INCR", width = "15px")
}
\arguments{
\item{root}{Root of the created input names. Names generated will be of the form "root___i___INCR", where i is the 1:n row number.}

\item{l}{Number of rows of the table.}

\item{buttonLabel}{Which text to display on the button. Default = "+"}

\item{stem}{Input suffix. Default = "INCR"}

\item{width}{The width of the input, e.g. '400px', or '100%'; see htmltools::validateCssUnit().\cr Default = "15px"}
}
\description{
A function to create a "+" button column fill-down button for use in DT data table columns.
Could be used for many different things depending on how you write the shiny server, here we use it for filling down with increment.
Based on shiny::actionButton.
}
\examples{
# Use within a shiny server as below:

lev <- unique(df$Choice)
width <- paste0(30*max(c(nchar(as.character(lev)), 2)), "px")
nr <- nrow(df)
server <- function(input, output, session) {

  Table <- df
  Table$Choice <- shinySelectInput(Table$Choice,
                                   "Choice",
                                   lev,
                                   width)
  Table$Choice___INCR <- shinyPlusFD("Choice",
                                     nr)

}

}
