% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPropInput.R
\name{shinyPropInput}
\alias{shinyPropInput}
\title{shinyPropInput}
\usage{
shinyPropInput(startVal, width = "300px", root = "Proportion")
}
\arguments{
\item{startVal}{Starting values of each proportion.}

\item{width}{The width of the input, e.g. "400px", or "100%"; see htmltools::validateCssUnit().}

\item{root}{Root of the created input names, default = "Proportion". Input names generated will be of the form "root___i", where i is the 1:n row number.}
}
\description{
Convenience wrapper for shiny::numericInput for use in DT data table columns.
This is specifically to query users for a proportion (between 0 and 1 by steps of 0.001).
}
\examples{
# Use within a shiny server as below:

server <- function(input, output, session) {

  Table$Proportion <- shinyPropInput(Table$Proportion)

}

}
