% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySelectInput.R
\name{shinySelectInput}
\alias{shinySelectInput}
\title{shinySelectInput}
\usage{
shinySelectInput(startVal, root, levels, width = "600px", NAs = TRUE)
}
\arguments{
\item{startVal}{Starting values of each text input.}

\item{root}{Root of the created input names. Names generated will be of the form "root___i", where i is the 1:n row number.}

\item{levels}{Allowed values.}

\item{width}{The width of the input, e.g. '400px', or '100%'; see htmltools::validateCssUnit().\cr Default = "600px"}

\item{NAs}{If TRUE (default), NA selections are allowed (and NA is included as another additional allowed level). If FALSE, they are automatically replaced with the first authorized level at initialization.}
}
\description{
Convenience wrapper for shiny::selectInput for use in DT data table columns.
}
\examples{
# Use within a shiny server as below:

lev <- unique(df$Choice)
width <- paste0(30*max(c(nchar(as.character(lev)), 2)), "px")
server <- function(input, output, session) {

  Table <- df
  Table$Choice <- shinySelectInput(Table$Choice,
                                   "Choice",
                                   lev,
                                   width)

}

}
