% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTextInput.R
\name{shinyTextInput}
\alias{shinyTextInput}
\title{shinyTextInput}
\usage{
shinyTextInput(startVal, root, width)
}
\arguments{
\item{startVal}{Starting values of each text input.}

\item{root}{Root of the created input names. Names generated will be of the form "root___i", where i is the 1:n row number.}

\item{width}{The width of the input, e.g. '400px', or '100%'; see htmltools::validateCssUnit().}
}
\description{
Convenience wrapper for shiny::textInput for use in DT data table columns.
}
\examples{
# Use within a shiny server as below:

server <- function(input, output, session) {

  Table$Free_input <- shinyTextInput("Enter something...",
                                     "FreeInput")

}

}
