% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.ID.R
\name{short.ID}
\alias{short.ID}
\title{short.ID}
\usage{
short.ID(Prot, Proteins.colname, Genes.colname)
}
\arguments{
\item{Prot}{The protein/protein groups file.}

\item{Proteins.colname}{The name of the protein IDs column.}

\item{Genes.colname}{The name of the gene IDs column.}
}
\description{
A convenient function to create a meaningful tag for proteins in graphs, using the gene and protein columns.
}
\examples{
PG$Short.ID <- function(PG, Proteins.colname = "Protein.IDs", Genes.colname = "Genes")

}
