% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_to_MQ.R
\name{Skyline_to_MQ}
\alias{Skyline_to_MQ}
\title{Skyline_to_MQ}
\usage{
Skyline_to_MQ(
  Skyline_fl,
  Fixed.mods = c("Carbamidomethyl"),
  N.clust,
  N.reserved = 1,
  cl,
  shinyOpt = "popup",
  digPattern = "KR"
)
}
\arguments{
\item{Skyline_fl}{External Skyline .tsv export file, which can be a PSMs- or Transitions-level table. If transitions, it will aggregate information by file, modified sequence and charge into PSM-style table.}

\item{Fixed.mods}{A vector of those Fixed modifications which will not be reported in the results (reported by diaNN, but MQ ignores them). Default = "Carbamidomethyl"}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{shinyOpt}{One of "popup" (default), "pane", "dialog", or "browser".}

\item{digPattern}{Default = "KR" for trypsin. Only used for the "Missed cleavages" column if that from Skyline is missing. Vector of amino acids after which the enzyme cleaves.}
}
\description{
Converts a Skyline .tsv or .csv export table (PSMs or Transitions) to a table with MaxQuant evidence.txt-like formatting.
As with the other functions of this type, the idea is not to get perfect conversion, but close enough that the data can be fed into this package's analysis scripts.

The output is a list of 2 items:
- Evidence: a data.frame similar to the evidence.txt table MaxQuant creates
- PTMs: a modifications table

Surprisingly, the diaNN PSMs table does not include MZ values! Is there a way to do this by adding a command line argument?
For now they are either detected from the library file (same folder) or computed.
Cysteine alkylation is assumed to be Carbamidomethyl
}
\examples{
temp <- Skyline_to_MQ(Skyline_fl = skyline_fl)
ev <- temp$Evidences
Modifs <- temp$PTMs

}
