% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds.R
\name{Thresholds}
\alias{Thresholds}
\title{Thresholds}
\usage{
Thresholds(
  values,
  ref = NULL,
  p = 0.01,
  Pvalue_threshold = 0.05,
  Distribution = "Cauchy"
)
}
\arguments{
\item{values}{The vector of distributed values for which we want a threshold.}

\item{ref}{The reference. If left as NULL, proportions are from the values vector. If provided as a vector of values, the function aims to find thresholds including a given proportion (based on modelled distributions) p of reference to values.}

\item{p}{The relevant proportion to include (either absolute, or contamination by a reference).}

\item{Pvalue_threshold}{Which threshold can we accept for the distribution test's P-value? Default = 0.05}

\item{Distribution}{Which Distribution does the data follow. Allowed values are "Normal" and "Cauchy" (default).}
}
\description{
Find thresholds to include only a given proportion of data.
The data has to be normalized as the function assumes symmetry of the problem relative to 0!!!
Unused, old, deprecated, candidate for deletion.
}
\examples{
Thresholds(values = test$Treated,
           ref = test$Control)

}
