% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.Profile.R
\name{Time.Profile}
\alias{Time.Profile}
\title{Time.Profile}
\usage{
Time.Profile(
  df,
  prot = "",
  prot.col = "Protein IDs",
  exprs.root = "Ratio_Mean log2_",
  exprs.name = "Ratio",
  error.root = FALSE,
  conds,
  tp,
  save = FALSE,
  title = "",
  labels = FALSE
)
}
\arguments{
\item{df}{The protein/protein groups file, which must include expression data (and SE)}

\item{prot}{Optional: the protein ID to be plotted. If left at its default value of "", all protein groups will be plotted on one graph.}

\item{prot.col}{The name of the Protein IDs columns. Default = "Protein IDs"}

\item{exprs.root}{The root of the names of expression/ratio columns. Default = "Ratio_Mean log2_"}

\item{exprs.name}{Optional. What should we call the expression dimension? Default = "Ratio"}

\item{error.root}{Optional argument (default = FALSE). The root of the names of error columns, if any are provided. It is probably not a good idea to provide it when plotting thousands of proteins.}

\item{conds}{The conditions to plot.}

\item{tp}{The time points to plot. This should be a named numeric (the name is what the function expects to find in the expression and error column names).}

\item{save}{Should the plot be saved? default = FALSE. Set it to a valid ggsave name to save it to the required format, e.g. "Time profiles.jpeg".}

\item{title}{Optional title (default = "").}

\item{labels}{default = FALSE. If set to a column name, will display these as protein labels to the left of each profile. Ignored when only plotting a single protein ID.}
}
\description{
A function to plot the time profile of protein(s).
The function assumes that expression column names follow the following pattern:
"root.cond.tp" (i.e. a root string then the condition then the time point).
If there are error columns, these should follow a similar pattern with their own root.
}
\examples{
# To plot a single protein's profile:
Time.Profiledf(df = PG, prot = "Q23098", prot.col = "Protein IDs",
                         exprs.root = "Ratio_Mean log2_ ", error.root = "SE.",
                         conds = Cond, tp = Time.Pts, save = FALSE)
# To plot the profiles of all the proteins in the data frame: 
Time.Profile(df = PG, prot.col = "Protein IDs",
                         exprs.root = "Ratio_Mean log2_", error.root = FALSE,
                         conds = Cond, tp = Time.Pts, save = FALSE)
}
