% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topattern.R
\name{topattern}
\alias{topattern}
\title{topattern}
\usage{
topattern(x, start = TRUE, end = FALSE, collapse = "|")
}
\arguments{
\item{x}{The character vector to be turned into a regex pattern.}

\item{start}{Default = TRUE, which means the pattern will require a match to the start of the string.}

\item{end}{Default = FALSE. Set to TRUE for the pattern require a match to the end of the string.}

\item{collapse}{For vectors of length > 1, character to collapse the vector to in order to create a single pattern. Default = "|". Set to FALSE to turn it off.}
}
\value{
A regex pattern.
}
\description{
A function to convert character strings to regex patterns
}
\details{
Internal function to convert a chain of characters into a regex pattern usable by gsub() or grep().
Note: this may not work for absolutely every string, but is a very useful shortcut for the controlled usage we make of it within this package.
}
\examples{
topattern(pep.ref[length(pep.ref)])

}
