% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopN.R
\name{TopN}
\alias{TopN}
\title{TopN}
\usage{
TopN(
  N = 3,
  Prot,
  Peptide.IDs = "Peptide IDs",
  Pep,
  id = "id",
  Pep.Intens.Nms,
  log.Pep.Intens = 10,
  Prot.TopN.to.log = 10,
  Mods,
  Mod.Nms = "Modified sequence",
  Min.Pep.Nb = 1,
  In.Norm = FALSE,
  Out.Norm = TRUE,
  Method = "mean",
  corr = "local"
)
}
\arguments{
\item{N}{How many peptides should be averaged?}

\item{Prot}{The protein/protein groups file.}

\item{Peptide.IDs}{The name of the Protein/Protein groups file's peptide IDs column. Default = "Peptide IDs"}

\item{Pep}{The peptides file.}

\item{id}{The name of the Peptides file IDs column. Default = "id"}

\item{Pep.Intens.Nms}{Name of the peptides intensity column(s)}

\item{log.Pep.Intens}{Set to 0 or FALSE if input peptide intensities are linear scale. If the data is already log scale, set to the relevant base. Default = 10}

\item{Prot.TopN.to.log}{Should the output protein TopN values be log-scale or not? If set to a number, this is assumed to be the base of the output log scale. Default = 10}

\item{Mods}{Which modifications should be included? If set to FALSE, will not filter any modifications.}

\item{Mod.Nms}{Default = "Modified sequence". The name of the column containing the Modified sequence in the peptides file. Can be set to a non-modified sequence if Mods = FALSE}

\item{Min.Pep.Nb}{How many peptides should at least be present? Should be at the very least 1 (set do default). If greater than N, will be ignored.}

\item{In.Norm}{Should the Input be re-normalised? Default = FALSE}

\item{Out.Norm}{Should the Output be re-normalised? Default = TRUE}

\item{Method}{Which averaging method should be used. The default, "mean", is an arithmetic mean on log values, equivalent to a geometric mean for linear values. Can be instead set to "median", or the name of any averaging function of choice which R knows.}

\item{corr}{Should cases where there are less than N peptides be corrected? Values can be "local" (default), "global" or FALSE. Local will do the correction for factors infered for each column. Global will average for all columns. FALSE will not do it. Ignored if N = 1}
}
\description{
A function to calculate protein Intensity-based TopN abundance estimator from individual peptide intensities.
}
\examples{
test <- TopN(N = 3, Prot = prot, Peptide.IDs = "Peptide IDs", Pep = pep, id = "id", Pep.Intens.Nms = Pep.Intensity.Names,
             log.Pep.Intens = 10, Prot.TopN.to.log = 10,
             Mods = Mod4Quant, Mod.Nms = "Modified sequence", Min.Pep.Nb = 1, In.Norm = TRUE, Out.Norm = TRUE, Method = "mean", corr = "local")
colnames(test) <- gsub("\\\\.Intensity\\\\.", ".", colnames(test))
prot[, colnames(test)] <- test

}
