% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volc.plot_2.R
\name{Volc.plot_2}
\alias{Volc.plot_2}
\title{Volc.plot_2}
\usage{
Volc.plot_2(
  data,
  ids,
  x,
  y,
  a = NULL,
  N,
  title,
  v.lines,
  h.lines,
  v.lines_min = NULL,
  labels = FALSE,
  displ.limit = 50,
  save = FALSE,
  print = TRUE,
  col_return = FALSE
)
}
\arguments{
\item{data}{The dataframe that contains the values.}

\item{ids}{The name of the column of ID values to display.}

\item{x}{The name of the column with values to display on the X axis.}

\item{y}{The name of the column with values to display on the Y axis.}

\item{a}{The optional name of the column which will be used to calculate the Alpha values of dots. NULL by default.}

\item{N}{Optional. The number of replicates that were used to calculate P values (Y axis).}

\item{title}{The title of the plot.}

\item{v.lines}{Vertical threshold (positive value: this is assumed to be symmetrical).}

\item{h.lines}{Horizontal thresholds.}

\item{v.lines_min}{Minimal value for vertical threshold. NULL by default.}

\item{labels}{Should labels be displayed (for dots beyond thresholds)? Default = FALSE.}

\item{displ.limit}{Maximum number of labels to display. Default = 50. Set to NULL to turn off.}

\item{save}{FALSE by default. Set to format you want pictures saved as.}

\item{print}{TRUE by default. But can be set to FALSE because sometimes you just want to save a plot for later without printing it.}

\item{col_return}{FALSE by default. Should a column of significance be added to the output?}
}
\description{
Another function for volcano plots, meant to be simpler than the main, awfully complex volcano.plot() function.
This has not been used for a long time so may prove unreliable.
}
\examples{
test <- Volc.plot_2(data = PG, ids = "Short ID", x = a1, y = a2,
                    N = length(grep(a1, colnames(PG), fixed = TRUE)), title = "Volcano plot",
                    v.lines = 2^v, h.lines = h)

}
