% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFasta.R
\name{writeFasta}
\alias{writeFasta}
\title{writeFasta}
\usage{
writeFasta(DB, destFl, filter, return = FALSE, newFl = FALSE)
}
\arguments{
\item{DB}{A data base of protein sequences. Must contain "Header" and "Sequence" column.}

\item{destFl}{Destination file path.}

\item{filter}{A character of protein accessions (IDs) to use as (positive) filter for the output. Requires that DB contain a "Protein ID" column.}

\item{return}{Should the resulting character string be returned (TRUE), or only written to destFl (default)?}

\item{newFl}{New file? If so the user will be prompted for the path of the destination file (and destFl will be ignored). Default = FALSE}
}
\description{
A function to write a data table of protein sequences, parsed from a fasta by Format.DB, as a fasta.
The fasta is:
 - written down if a destination file is provided using destFl,
 - or, if newFl is TRUE, a modal dialog is opened to interactively choose it.
 - The fasta can also optionally be returned as a character string.
}
\examples{
DB <- Format.DB(".../My fasta database.fasta")
path <- ".../My fasta database's backup.fasta"
writeFasta(DB, path)
fasta <- writeFasta(DB, return = TRUE) # Assign as object in environment without writing to file.
fasta <- writeFasta(DB, path, return = TRUE) # Write to file & assign as object in environment.
# Filter
filt <- c("Protein 1", "Protein 2")
path <- ".../My filtered fasta database.fasta"
writeFasta(DB, path, filt)

}
