% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_bias.R
\name{AA_bias}
\alias{AA_bias}
\title{AA_bias}
\usage{
AA_bias(Ev, DB, unique = FALSE)
}
\arguments{
\item{Ev}{A data frame containing a "Sequence" column of unmodified peptide sequences.}

\item{DB}{The reference proteome, a data.frame containing a "Sequence" column of unmodified protein sequences.}

\item{unique}{Logical, default = FALSE. If TRUE, we will look at unique sequences (not observations).}
}
\value{
A table of amino acid representation statistics for the current dataset.
}
\description{
A little function to test for biases against amino acids in the current (observed) dataset, versus the parent proteome.
}
\examples{
AA.bias <- AA_bias(Ev = ev, DB = db)

}
