% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.log.R
\name{diff.log}
\alias{diff.log}
\title{diff.log}
\usage{
\method{diff}{log}(p, dat)
}
\arguments{
\item{p}{Starting estimates.}

\item{dat}{Data frame of log-transformed, row-normalized quantitative profiles.}
}
\description{
A function whose sole purpose is to be turned into another whose sum will be minimized by nls.lm within the LFQ function.
}
\examples{
temp3 <- ... # Data frame of log-transformed, row-normalized quantitative values
diff.log.v <- function(...) {
  p <- list(...)
  res <- diff.log(p, dat = temp3)
  return(res)
}

}
