% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_ratio_av.R
\name{log_ratio_av}
\alias{log_ratio_av}
\title{log_ratio_av}
\usage{
log_ratio_av(x)
}
\arguments{
\item{x}{Log-transformed ratio values to be averaged.}
}
\description{
Averages log-transformed ratio values.\cr
 - The mean is used when some valid values are present\cr
 - If only infinite values are present, if these are all the same (-Inf or +Inf, not both), then this consensus value is returned, otherwise NA is returned\cr
 - All NA or NaN values return NA\cr
}
