% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RefRat.R
\name{make_RefRat}
\alias{make_RefRat}
\title{make_RefRat}
\usage{
make_RefRat(
  data = pep,
  experiment.map = Exp.map,
  experiment.map_col = "Ref.Sample.Aggregate",
  int.root = pep.ref[length(pep.ref)],
  rat.root = pep.ratios.ref[1],
  rat.con.grps = RatConGrps,
  mode = RefRat_Mode,
  parameters = Param,
  logInt = FALSE,
  logRat = 2
)
}
\arguments{
\item{data}{A data.frame containing expression values. Default = pep}

\item{experiment.map}{The experiment map. Default = Exp.map}

\item{experiment.map_col}{Name of the sample column in the experiment map.}

\item{int.root}{Name root for expression columns. Default = pep.ref[length(pep.ref)]}

\item{rat.root}{Name root for ratio columns. Default = pep.ratios.ref[1]}

\item{rat.con.grps}{Which sample groups are we considering. Default = RatConGrps}

\item{mode}{One of 1 or 2}

\item{parameters}{The experiment's parameters file. Default = Param}

\item{logInt}{Set to 0 or FALSE if input peptide intensities are linear scale. If the data is already log scale, set to the relevant scale's base. Default = FALSE}

\item{log.Ratios}{Set to 0 or FALSE if input peptide ratios are linear scale. If the data is already log scale, set to the relevant scale's base. Default = 2}
}
\value{
Reference sample-to-sample log fold changes.
}
\description{
A function to create reference ratios according to a variety of methods, for the purpose of a defining logFC thresholds.
}
\details{
This function can be used as a logFC-level filter for DEPs - after statistical analysis.
Instead of using a fixed logFC threshold, this can instead be used to define thresholds based on internal data variability.
}
\examples{
pep.Ref.Ratios \%<o\% make_RefRat()
pep[, colnames(pep.Ref.Ratios)] <- pep.Ref.Ratios              

}
