% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMet_LCMS.R
\name{MatMet_LCMS}
\alias{MatMet_LCMS}
\title{MatMet_LCMS}
\usage{
MatMet_LCMS(
  ScanHdsMnLoc = "C:/ScanHeadsman-1.2.20200730",
  LocalRoot = "D:/groups_temp/",
  ArchiveRoot = "Q:/MS/Acquired data/",
  RawFiles = rawFiles,
  Columns = "default"
)
}
\arguments{
\item{ScanHdsMnLoc}{ScanHeadsman path.\cr Default = "C:/ScanHeadsman-1.2.20200730" - so far for us version 1.2 works, 1.3 does not.}

\item{LocalRoot}{What is the root of the folder structure in the local folder. If the raw files are not local but were moved to an archive, this root will be replaced by that of the archive, see ArchiveRoot.\cr Default = "D:/groups_temp/"}

\item{ArchiveRoot}{If the raw files are not local but were moved to an archive, where would that be?\cr Default = "Q:/MS/Acquired data/"}

\item{RawFiles}{The raw files from which to attempt to extract the method.}

\item{Columns}{Path to a table of LC columns to pick from.}
}
\description{
A function to write the LC-MS part of Materials & Methods, pulling information from the methods file.
For now only supports DIA methods from a Q-Exactive type instrument.

!!! When editing function, always save with UTF-8 encoding!!!
}
\examples{
MatMet_Text2 <- MatMet_LCMS()

}
