% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.Param.aggreg.R
\name{parse.Param.aggreg}
\alias{parse.Param.aggreg}
\title{parse.Param.aggreg}
\usage{
parse.Param.aggreg(
  parameter,
  aggregates = Aggregates,
  map = Aggregate.map,
  aggr.list = Aggregate.list
)
}
\arguments{
\item{parameter}{The parameter to parse.}

\item{aggregates}{The annotations (basic aggregate building blocks) as a named vector. Default: Aggregates}

\item{map}{The reference aggregates map. All existing aggregates in this map must exist. Default: Aggregate.map}

\item{aggr.list}{The list of aggregates. Default: Aggregate.list}
}
\description{
A function to parse an aggregates-type parameter.
Only use for assignments to temporary variables, otherwise use the newer parse.Param.aggreg.2
}
\examples{
Ratios.Plot.split <- parse.Param.aggreg(parameter = Param$Ratios.Plot.split)

}
