% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEval.R
\name{plotEval}
\alias{plotEval}
\title{plotEval}
\usage{
plotEval(plot)
}
\arguments{
\item{plot}{ggplot to evaluate.}
}
\value{
Evaluated ggplot.
}
\description{
A simple wrapup for evaluation a ggplot, allowing to save it in a parallel manner.
}
