% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtMatch.R
\name{ProtMatch}
\alias{ProtMatch}
\title{ProtMatch}
\usage{
ProtMatch(
  Seq,
  DB,
  Cut = c("K_", "R_"),
  strict.avoid = "",
  loose.avoid = c("K_P", "R_P"),
  min = 7,
  max = 100,
  missed = 2,
  I_eq_L = TRUE
)
}
\arguments{
\item{Seq}{Character vector of peptide sequences.}

\item{DB}{The formated protein database, with proteins IDs and sequences.}

\item{Cut}{For digestion. The character vector of amino acid(s) after which to cut. For each, provide one letter followed or preceded by underscore indicating whether the cut is C- or N-terminal, respectively. Default (for Trypsin) = c("K_", "R_")}

\item{strict.avoid}{For digestion. The pattern(s) that will strictly block cleavage when overlapping with cleavage sites. Write cleavage site as underscore. Default = ""}

\item{loose.avoid}{For digestion. The pattern(s) that will partially block cleavage when overlapping with cleavage sites. Write cleavage site as underscore. Default =  c("K_P", "R_P")}

\item{min}{For digestion. The minimum length of peptides to report. Default = 7}

\item{max}{For digestion. The maximum length of peptides to report. Set to 100 (default) to allow even very large peptides. If one of the peptides in Seq in longer, than that length will be used instead.}

\item{missed}{Let's say we have observations with up to 6 missed cleavages. This will probably break the function. In this case, it is better to specify a smaller number of missed cleavages, e.g. 2.}

\item{I_eq_L}{Should we consider I and L identical? Currently, by default, TRUE for both DIA and DDA: see https://github.com/vdemichev/DiaNN/discussions/1631}
}
\description{
This function is very slow and should not be used! It is only provided for legacy purposes. Instead use ProtMatch2, which is ~50-to-1000 times faster!

Some search software do not report all proteins from the database which match a specific sequence.
Others, such as MaxQuant, do report weird matches for a minority of PSMs, or miss a few perfectly valid matches.
While this may be correct from their point of view, this function allows re-checking matches for consistency between software.
This will return a data frame of matches between observed peptides and protein IDs from a formatted protein sequences database.
}
