% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFDInput.R
\name{shinyFDInput}
\alias{shinyFDInput}
\title{shinyFDInput}
\usage{
shinyFDInput(root, l, useIcon = FALSE, width = "15px")
}
\arguments{
\item{root}{Root of the created input names. Names generated will be of the form "root___i___FD", where i is the 1:n row number.}

\item{l}{Number of rows of the table.}

\item{useIcon}{Ignored, only kept because of old code!}

\item{width}{The width of the input, e.g. "400px", or "100%"; see htmltools::validateCssUnit().\cr Default = "15px"}
}
\description{
A function to create column fill-down buttons for use in DT data table columns.
Based on shiny::actionButton.
}
\examples{
# Use within a shiny server as below:

lev <- unique(df$Choice)
width <- paste0(30*max(c(nchar(as.character(lev)), 2)), "px")
nr <- nrow(df)
server <- function(input, output, session) {

  Table <- df
  Table$Choice <- shinySelectInput(Table$Choice,
                                   "Choice",
                                   lev,
                                   width)
  Table$Choice___FD <- shinyFDInput("Choice",
                                    nr,
                                    TRUE)

}

}
