% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano.plot.R
\name{Volcano.plot}
\alias{Volcano.plot}
\title{Volcano.plot}
\usage{
Volcano.plot(
  Prot,
  mode,
  X.root,
  Y.root,
  X.normalized = TRUE,
  experiments.map,
  aggregate.map,
  aggregate.name,
  aggregate.list,
  parameters,
  save = FALSE,
  labels = "both",
  MaxLabels = 50L,
  MaxLabels_priority = "X",
  Ref.Ratio.values = NULL,
  Ref.Ratio.method = "obs2",
  ratios.FDR = 0.01,
  FDR.thresh,
  FDR.root = "Significant-FDR=",
  arbitrary.lines = NULL,
  arbitrary.thresh = NULL,
  proteins = NULL,
  proteins_split = FALSE,
  IDs.col = "Protein IDs",
  Proteins.col = "Protein IDs",
  return = FALSE,
  return.plot = FALSE,
  title = "",
  title.root = "",
  subfolder = "Reg. analysis/t-tests",
  subfolderpertype = TRUE,
  Symmetrical,
  Alpha = 1L,
  Alpha.identity = FALSE,
  Alpha.labels = FALSE,
  Alpha.max = 1L,
  Alpha.min = 0.01,
  Size = 1L,
  Size.max = 3L,
  Size.min = 0.01,
  cex = 2L,
  lineheight = 1L,
  Xlim = Inf,
  Ylim = Inf,
  plotly = FALSE,
  plotly_local = TRUE,
  plotly_username,
  plotly_API_key,
  plotly_subfolder = "",
  plotly_sharing = "secret",
  plotly_labels,
  Contaminants = FALSE,
  N.clust,
  N.reserved = 1L,
  cl,
  reg.root,
  SAM = FALSE,
  curved_Thresh,
  saveData = FALSE,
  contrasts,
  X.root_ind,
  X.root_ind_base = 10L,
  show.labels = TRUE
)
}
\arguments{
\item{Prot}{The protein/protein groups table. It will be added columns saying whether a protein group is regulated or not for each test performed (i.e. each "aggregate")}

\item{mode}{One of "standard" or "custom". Ignored if SAM = TRUE.}

\item{X.root}{Root of the names for X-axis log-scale values. If missing, defaults are provided in mode = "custom" in the parameters file.}

\item{Y.root}{Root of the names for Y-axis values (usually -log10(Pvalues)). If missing, default are provided in mode = "custom" in the parameters file.}

\item{X.normalized}{Logical. Is X normalized? Assumed to be TRUE by default.}

\item{experiments.map}{The data.frame defining the structure of the dataset.}

\item{aggregate.map}{The aggregate map.}

\item{aggregate.name}{The name of the aggregate to use.}

\item{aggregate.list}{The named list of aggregates.}

\item{parameters}{The experiment's parameters file.}

\item{save}{Logical or character. Should the plot(s) be saved? Default = FALSE. Set it to a vector of ggplot2-compatible file extensions to save to the corresponding format.}

\item{labels}{How to decide which point labels will be printed. Values are:\cr
- "proteins", in which case parameters MaxLabels and MaxLabels_priority will also be used;\cr
- "thresholds", in which case proteins above thresholds will be labelled;\cr
- "FDR", in which case only points that pass Benjamini-Hochberg FDR (or are in the list of proteins of interest) will be displayed.\cr
- The default, "both" (default), which means "proteins + thresholds".}

\item{MaxLabels}{Maximum number of labels to print. Does not apply to matches to the list provided by "proteins". Default = 50}

\item{MaxLabels_priority}{If not printing all labels, will we favour more extreme ratios (default = "X") or more significant values ("Y")?}

\item{Ref.Ratio.values}{Used if the default parameter values for Ratio thresholds are to be replaced with dynamic, reference-based ones. Also requires a few arguments below.}

\item{Ref.Ratio.method}{Defines how vertical ratios are computed.\cr
- "obs2": current default, will sort observations within sample groups and identify a threshold such that a "ratios.FDR" proportion of most extreme values are selected.\cr
- "obs1": does almost the same as the above, but only for control-to-control or control-to-average-control ratios within the ratios group.\cr
- "SD": same as the previous, but will calculate the threshold using qnorm(ratios.FDR, median(Ref.Ratio.values), sd(Ref.Ratio.values)), i.e. assuming normality.\cr}

\item{ratios.FDR}{The proportion of reference ratios to be accepted, if Ref.Ratio.values is provided. Default = 0.01.}

\item{FDR.thresh}{The significance thresholds calculated for each acceptable FDR value.}

\item{FDR.root}{The root of the columns with Significance for different FDRs. Default = "Significant-FDR="}

\item{arbitrary.lines}{Default = NULL. Data.frame of arbitrary lines to add to the graph. Must have 5 columns: "slope", "yintercept", "xintercept" (if vertical line, else NA), "colour", "label".}

\item{arbitrary.thresh}{Default = NULL, ignored if mode = FDR. Similar to the above, but will define only horizontal lines. The lowest will determine the threshold below which no labels are displayed (except when using the "proteins" argument).}

\item{proteins}{Optional, the character list of protein IDs to be highlighted.}

\item{proteins_split}{Logical. Default = FALSE. If TRUE, and if there are protein matching the provided list (see "proteins" argument), then these will be highlighted in a separate graph.}

\item{IDs.col}{Default = "Protein IDs".}

\item{Proteins.col}{Default = "Protein IDs"; used for identifying matches to the "proteins" argument. IF missing, IDs.col will be used.}

\item{return}{Logical. Should we return a column with significance levels? Default = FALSE.}

\item{return.plot}{Logical. Should we return the plot itself? Default = FALSE.}

\item{title}{It is possible to set a specific title. Failing that a default one will be provided based on the value of the mode argument.}

\item{title.root}{Optional argument to add a character string at the beginning of the file names when saving.}

\item{subfolder}{Name of the sub-folder where graphs are to be saved. Default = "Reg. analysis/t-tests"}

\item{subfolderpertype}{Logical. If TRUE (default), will save each type of graph in a dedicated sub-folder.}

\item{Symmetrical}{Are we also interested in the left half (down-regulated) or only the right one (up-regulated)? May be used to override the symmetry provided by contrasts. If it and contrasts are missing, and "parameters" is provided, the function will attempt to infer it from parameter "Type".}

\item{Alpha}{Either a single numeric value, or a length 1 character: the name of a numeric column in Prot mapping each dot's alpha level to a column. Default = 1 (no alpha variations)}

\item{Alpha.identity}{Logical. Do we need to fit the Alphas into a 0:1 scale? Default = FALSE}

\item{Alpha.labels}{Logical. Should labels also be affected by alpha? Default = FALSE}

\item{Alpha.max}{For when mapping alpha to a column. The highest allowed alpha. Default = 1}

\item{Alpha.min}{For when mapping alpha to a column. The weakest allowed alpha. Default = 0.01}

\item{Size}{Either a single numeric value, or a length 1 character: the name of a numeric column in Prot. Default = 1 (same size for all)}

\item{Size.max}{For when mapping size to a column. The largest allowed dot size. Default = 2}

\item{Size.min}{For when mapping size to a column. The smallest allowed dot size. Default = 0.01}

\item{cex}{Label text size. Default = 2}

\item{lineheight}{Label interline size. Default = 1}

\item{Xlim}{Set fixed limits for the X axis.}

\item{Ylim}{Set fixed limits for the Y axis.}

\item{plotly}{Logical. Should we generate a plotly? FALSE by default; if TRUE, the link will be automatically returned.}

\item{plotly_local}{Deprecated (now always TRUE)}

\item{plotly_username}{Deprecated}

\item{plotly_API_key}{Deprecated}

\item{plotly_subfolder}{Default = "". Name of the plotly server sub-folder where plotly graphs are to be saved.}

\item{plotly_sharing}{Deprecated}

\item{plotly_labels}{Labels to be displayed in the plotly tooltip. A character corresponding to valid names of Prot columns. Naming this vector provides a way to override the original column names: what will be displayed in the tooltip will be the name in this vector, not the column name itself.}

\item{Contaminants}{Logical. If FALSE (default), contaminants will be filtered out prior to drawing the plot(s).}

\item{N.clust}{A limit on the number of vCPUs to use. If left as NULL (default), uses the number of available clusters - 1, to a minimum of 1.}

\item{N.reserved}{Default = 1. Number of reserved vCPUs the function is not to use. Note that for obvious reasons it will always use at least one.}

\item{cl}{Already have a cluster handy? Why spend time making a new one, which on top of that may invalidate the old one. Just pass it along!}

\item{reg.root}{Classically, this function would output a "Regulated - ..." column per sample group, based on applying the thresholds (either provided directly - significance is read from the table using "FDR.root", and FC threshold are either applied directly using "arbitrary.thresh" or calculated based on "Ref.Ratio.values"). Providing this input allows the function to bypass this step and apply colors based on a pre-determined decision (used for the F-test currently). Automatically set to "Regulated - " if SAM is TRUE.}

\item{SAM}{Logical. If TRUE, then a curved threshold is applied based on the SAM analysis. FALSE by default.}

\item{curved_Thresh}{Curved threshold parameters, only used if mode = "curved".\cr A named list, with one item for each (same names as the ones appended to X.root and Y.root in Prot column names).\cr Each item should be a named list of S0 (single value), Si (single value) and d (one value per FDR level) numerics.}

\item{saveData}{Logical. If TRUE, a file containing the processed long-format data used to create each plot will be saved locally. Default = FALSE.}

\item{contrasts}{If provided, supersedes experiments.map for the purpose of detecting samples (through the "Contrast" column), and if containing the "Up-only" column also defines whether plots are two-sided or not.}

\item{X.root_ind}{Optional root of the column names of individual expression values, which can then be displayed in the tooltip.}

\item{X.root_ind_base}{Log base of individual expression values. Default = 10L}

\item{show.labels}{Currently unused (deprecated).}
}
\value{
A list with at least one element ("Thresholds"). If return == TRUE, the Protein_groups_file - which may contain additional Regulated columns. If return.plot == TRUE, the evaluated ggplots. If plotly == TRUE, the plotly plots.
}
\description{
A function to create volcano plots.
}
\details{
This monster of a function can draw volcano plots, but can also be used for the decision on regulation.
}
\examples{
                
PG <- Volcano.plot(Prot = PG, mode = "custom", experiments.map = Exp.map,
                   aggregate.map = Aggregate.map,
                   aggregate.name = Volcano.plots.Aggregate.Level$aggregate,
                   parameters = Param,
                   save = c("jpeg", "pdf"),
                   labels = "both",
                   proteins = prot.list,
                   return = TRUE,
                   title.root = "FDR-type ")

}
